### 2019-12-06 #1937 @whedon
[PRE REVIEW]: ggalluvial: Layered Grammar for Alluvial Plots
**Submitting author:** @corybrunson (<a href="http://orcid.org/0000-0003-3126-9494">Jason Cory Brunson</a>)
**Repository:** <a href="https://github.com/corybrunson/ggalluvial" target ="_blank">https://github.com/corybrunson/ggalluvial</a>
**Version:** v0.11.1
