### 2019-12-06 #1937 @danielskatz
👋 @corybrunson - this submission needs a license - please add a plain text LICENSE file with the OSI-approved license that you are using for the software (which I think is GPL-3).
