### 2019-12-06 #1937 @danielskatz
Also, please check to see if the missing DOI whedon suggests is correct, and if so, add it to the .bib file. (Whedon suggestions are just suggestions, and sometimes they are wrong.)
