### 2019-12-06 #1937 @corybrunson
@danielskatz thank you! I've added the missing DOI. To be sure, is it good practice to simply put the standard GPLv3 license text (file name `LICENSE`) in the root directory, assuming i don't intend any deviations from that standard?
