### 2019-12-06 7f6bb79
rm file fields + add wilkinson doi in paper bib
