+++ paper.bib
@@ -34,7 +34,6 @@ abstract = {This article provides an illustrated history of the visual and conce
 author = {Friendly, Michael},
 booktitle = {Journal of Computational and Graphical Statistics},
 doi = {10.1198/106186002317375631},
-file = {:Users/corybrunson/Documents/Mendeley Desktop/Friendly - 2002 - A Brief History of the Mosaic Display.pdf:pdf;:Users/corybrunson/Documents/Mendeley Desktop/Friendly - 2002 - A Brief History of the Mosaic Display(2).pdf:pdf},
 issn = {10618600},
 keywords = {Cartogram,Data visualization,Log-linear models,Mosaic matrix,Space-filling displays,Thematic cartography,Tree map},
 number = {1},
@@ -49,7 +48,6 @@ abstract = {We present a system that allows users to interactively explore compl
 author = {Riehmann, P. and Hanfler, M. and Froehlich, B.},
 booktitle = {IEEE Symposium on Information Visualization, 2005. INFOVIS 2005.},
 doi = {10.1109/INFVIS.2005.1532152},
-file = {:Users/corybrunson/Documents/Mendeley Desktop/Riehmann, Hanfler, Froehlich - 2005 - Interactive Sankey diagrams.pdf:pdf},
 isbn = {0-7803-9464-X},
 pages = {233--240},
 publisher = {IEEE},
@@ -63,16 +61,16 @@ address = {Berlin/Heidelberg},
 author = {Friendly, Michael},
 booktitle = {Classification — the Ubiquitous Challenge. Studies in Classification, Data Analysis, and Knowledge Organization},
 doi = {10.1007/3-540-28084-7_4},
-file = {:Users/corybrunson/Documents/Mendeley Desktop/Friendly - 2005 - Milestones in the History of Data Visualization A Case Study in Statistical Historiography.pdf:pdf;:Users/corybrunson/Documents/Mendeley Desktop/Friendly - 2005 - Milestones in the History of Data Visualization A Case Study in Statistical Historiography(2).pdf:pdf},
 pages = {34--52},
 publisher = {Springer-Verlag},
 title = {{Milestones in the History of Data Visualization: A Case Study in Statistical Historiography}},
-url = {http://link.springer.com/10.1007/3-540-28084-7{\_}4 http://www.math.usu.edu/{~}symanzik/teaching/2009{\_}stat6560/Downloads/Friendly{\_}milestone.pdf},
+url = {http://www.math.usu.edu/{~}symanzik/teaching/2009{\_}stat6560/Downloads/Friendly{\_}milestone.pdf},
 year = {2005}
 }
 @book{Wilkinson2006,
 author = {Wilkinson, Leland},
 edition = {2},
+doi = {10.1007/0-387-28695-0},
 isbn = {0387286950},
 pages = {691},
 publisher = {Springer Science {\&} Business Media},
@@ -85,7 +83,6 @@ abstract = {Categorical data dimensions appear in many real-world data sets, but
 author = {Kosara, Robert and Bendix, Fabian and Hauser, Helwig},
 booktitle = {IEEE Transactions on Visualization and Computer Graphics},
 doi = {10.1109/TVCG.2006.76},
-file = {:Users/corybrunson/Documents/Mendeley Desktop/Kosara, Bendix, Hauser - 2006 - Parallel sets Interactive exploration and visual analysis of categorical data.pdf:pdf},
 issn = {10772626},
 keywords = {Categorical data,Information visualization,Interaction,Multivariate data,Nominal data},
 month = {jul},
@@ -100,7 +97,6 @@ year = {2006}
[…]
