### 2019-12-06 #1937 @danielskatz
> To be sure, is it good practice to simply put the standard GPLv3 license text (file name LICENSE) in the root directory, assuming i don't intend any deviations from that standard?

Yes, and if there are any deviations, you have changed the license to a non-OSI-approved one, which would mean that JOSS would not consider the submission.
