### 2019-12-06 #1937 @corybrunson
> Yes, and if there are any deviations, you have changed the license to a non-OSI-approved one, which would mean that JOSS would not consider the submission.

Right, thank you for pointing that out. In a moment i'll push the Markdown GPLv3 obtained from [here](https://www.gnu.org/licenses/gpl-3.0.en.html), renamed `LICENSE.md`.
