### 2019-12-06 #1937 @danielskatz
not LICENSE.md but LICENSE
