### 2019-12-06 #1937 @corybrunson
> not LICENSE.md but LICENSE

Thank you, this has been changed (using the plain text version of GPLv3).
