### 2019-12-19 fe3f459
resize image for joss paper pdf
