### 2019-12-22 a9c0f28
expand only function to method - preserves factor levels when returns na
