### 2019-12-22 e0c9a3b
restore only to standalone function + use switch for variable types
