### 2020-01-01 bcccfd5
typos in paper
