+++ paper.md
@@ -86,7 +86,7 @@ While most uses might be served equally well by other parallel sets plots or San
 [^opinion]: To be sure, this is a subjective assessment that may be refuted by visualization effectiveness research.
 
 **Repeated ordinal measures data:**
-Most Sankey, parallel sets, and alluvial implementations stack each bar plot in order of name or of size, though some follow user-provided hierarchies, and most insert gaps between categories for easy visual discrimination. Ordinal variables are most appropriately stacked in their own intrinsic and consistent order, and, when the number of categories changes from axis to axis, vertical separations can obscure whether magnitude totals changed as well. A use case by @Schlotter2019, to represent changes (or not) in patients' physical limitations following an investigational right heart valve repair technique, illustrates the use of an ordinal stratum variable (a heart failure functional classification). Another, by @North2019, to represent ranked preferences among several definitions of veganism by survey respondents, illustrates the importance of consistency in their order. In both cases, the fixed heights of the bar plots conveyed that no individuals were lost to follow-up.
+Most Sankey, parallel sets, and alluvial implementations stack each bar plot in order of name or of size, though some follow user-provided hierarchies, and most insert gaps between categories for easy visual discrimination. Ordinal variables are most appropriately stacked in their own intrinsic and consistent order and, when the number of categories changes from axis to axis, vertical separations can obscure whether magnitude totals changed as well. A use case by @Schlotter2019, to represent changes (or not) in patients' physical limitations following an investigational right heart valve repair technique, illustrates the use of an ordinal stratum variable (a heart failure functional classification). Another, by @North2019, to represent ranked preferences among several definitions of veganism by survey respondents, illustrates the importance of consistency in their order. In both cases, the fixed heights of the bar plots conveyed that no individuals were lost to follow-up.
 
 **Incomplete longitudinal data:**
 Alluvial plots clearly indicate times at which longitudinal data are censored or otherwise missing: Certain strata, or the alluvia or flows connecting them, are present at one time point but absent at a previous or future one.
@@ -94,10 +94,10 @@ Alluvial plots clearly indicate times at which longitudinal data are censored or
 @Sjoding2019 use a similar plot to trace patient groups receiving mechanical ventilation based on discretized tidal volumes, including a grey stratum for patients discontinued from intubation.
 
 **Signed categorical data:**
-@Edwards2019 produced a novel alluvial plot to represent changes in ownership category of owners in a halibut fishery. The total number of owners changed from year to year as exeters were not exactly matched by new entrants. In order to depict an accurate total but include both new entrants and exeters at each year, the authors affixed a negative stratum for the exeter category to each bar plot, reflected across the dimensional axis. Such a feature has no analogue in Sankey diagrams or parallel sets plots but potentially wide-ranging applications: Bar plots may use "positive" and "negative" bars to represent definitionally signed categories, such as contributors to revenue versus deficit, or to contrast the bars divided along a binary variable such as gender across age groups in a population ("pyramid plots"). Alluvial plots provide a way to track cases and cohorts across such graphics, even if cases change sign. Future use cases may demonstrate the practical utility of this functionality.
[…]
