### 2020-01-03 3b95595
debug self-adjoin for continuous x values + stat-stratum example
