### 2020-01-03 b9fb71e
clarify data documentation
