### 2020-01-06 02c3346
improvements + alternatives to splines
