### 2020-01-16 #2017 @whedon
[REVIEW]: ggalluvial: Layered Grammar for Alluvial Plots
**Submitting author:** @corybrunson (<a href="http://orcid.org/0000-0003-3126-9494">Jason Cory Brunson</a>)
**Repository:** <a href="https://github.com/corybrunson/ggalluvial" target ="_blank">https://github.com/corybrunson/ggalluvial</a>
**Version:** v0.11.3
