### 2020-01-17 #2017 @clauswilke
Overall this looks good to me. I have a few minor comments:
- Consistent use of lower-case "i" looks strange to me. I think the writing should follow the standard English convention of using upper-case "I" when referring to oneself.
- The first example in the (README)[https://github.com/corybrunson/ggalluvial/blob/master/README.md] appears to use a deprecated parameter.
- I would like to ask the author to put a little more effort into comparing this package to other approaches. In particular, ggforce has a very mature parallel sets implementation, and it is not clear to me how the ggforce features are similar or different from the ggalluvial features. This would benefit from a little more exposition. Also, while I understand that Sankey plots are very different from the plots generated by ggalluvial, I don't understand the difference between the alluvial plots and parallel sets. Is it just the gaps between groupings? In this context, I'd like to point out that ggforce puts gaps between groupings, but those gaps can be disabled as far as I remember.

Finally, in the comparison to other packages, I don't think it's so important to create a contrast between the present work and other works. This feels forced in places. It's useful to have multiple high-quality implementations of similar concepts, and there is no need to create an artificial contrast or show that one is better than the other. Just state what ggalluvial does and doesn't do, and what other packages do and don't do.
