### 2020-01-17 f5b43a9
replace deprecated param from readme
