### 2020-01-17 ece5102
revise & extend comparisons with other packages
