+++ paper.rmd
@@ -44,7 +44,7 @@ Second, ggalluvial adopts a distinctive geological nomenclature to distinguish "
 
 ## Functionality
 
-[The titular vignette](http://corybrunson.github.io/ggalluvial/articles/ggalluvial.html) thoroughly describes and illustrates the functionality of ggalluvial, and the reader is encouraged to browse [the package documentation](http://corybrunson.github.io/ggalluvial/reference/index.html) for comprehensive examples. In brief, the package contains stat and geom functions to add the following layers to a ggplot2 object:
+[The primary vignette](http://corybrunson.github.io/ggalluvial/articles/ggalluvial.html) thoroughly describes and illustrates the functionality of ggalluvial, and the reader is encouraged to browse [the package documentation](http://corybrunson.github.io/ggalluvial/reference/index.html) for comprehensive examples. In brief, the package contains stat and geom functions to add the following layers to a ggplot2 object:
 
 * _strata_, or stacked bar plots, located in parallel along a (plotting) axis of (variable) _axes_ or _dimensions_
 * _alluvia_, ribbons through strata that connect the categories of individual cases or cohorts at different axes
@@ -104,8 +104,6 @@ The alluvial stats require custom aesthetics---either `stratum` and/or `alluvium
 
 Most of the stat parameters control how the strata at each axis, and the lodes within each stratum, are ordered vertically. By default, these orderings are independent of differentiation aesthetics, so that layers are consistent within and across plots unless otherwise specified. [An auxiliary vignette](http://corybrunson.github.io/ggalluvial/articles/order-rectangles.html) details the effects of these parameters. They can also be set as global options.
 
-While questions and requests by users have prompted many improvements, the current version is feature complete with one exception: improvements and alternatives to the flow elements connecting lodes. The current splines will remain the default.
-
 ## Concepts
 
 Visualizations of flow processes have long encoded magnitudes as ribbon widths, constituting a type called Sankey diagrams [@Schmidt2008].
@@ -114,17 +112,31 @@ Their design anticipated parallel sets plots, which were adapted from parallel c
 These in turn anticipated "alluvial diagrams", proposed to visualize changes in case memberships across successive cross-sections [@Rosvall2010].
 Several R packages have been developed to generate these diagrams, in particular alluvial [@Bojanowski2016], ggparallel [@Hofmann2013], ggforce [@Pedersen2019], ggalluvial [@Brunson2019], and ggpcp [@Ge2019].
 
-The diagrams produced by ggalluvial are statistical graphics in the sense that they communicate statistical information using graphical methods [@Friendly2005] and, more narrowly, are uniquely determined from data by a fixed set of plotting rules [@Wilkinson2006]. The package adopts for them the term _alluvial plot_.[^plots]
-These plots are distinct from Sankey diagrams (they do not necessarily represent flow) and constitute a subtype of parallel sets plots distinguished by two features: a strict order on the stacked elements at each axis, including both the values of the discrete variables and the ribbons connecting cases or cohorts between them; and a meaningful positional dimension along which these elements are stacked, which precludes inserting gaps between them, features shared only by ggparallel.
-The package includes several levers of control over the statistical transformations, and thereby over the messages conveyed by the plot.[^easy]
-The plot elements are rendered by separate layers, following the additive (`+`) syntax of ggplot2, a feature shared only by ggpcp.
-
-[^plots]: This has the unfortunate side effect of conflating search results from the geology literature.
-[^easy]: Other ggplot2 extensions are not so flexible; in particular, the dependency package easyalluvial [@Koneswarakantha2019] was built on top of ggalluvial to exchange much of this flexibility for more expedient data exploration.
-
 Sankey, parallel sets, and alluvial diagrams are often conflated, and there is currently no consensus on what features are distinctive to each type.
[…]
