+++ paper.bib
@@ -78,6 +78,11 @@ title = {{The Grammar of Graphics}},
 url = {https://books.google.com/books?id=NRyGnjeNKJIC},
 year = {2006}
 }
+@misc{Slowikowski2019,
+author = {Slowikowski, Kamil},
+title = {{ggrepel: Automatically Position Non-Overlapping Text Labels with 'ggplot2'}},
+year = {2019}
+}
 @inproceedings{Kosara2006,
 abstract = {Categorical data dimensions appear in many real-world data sets, but few visualization methods exist that properly deal with them. Parallel Sets are a new method for the visualization and interactive exploration of categorical data that shows data frequencies instead of the individual data points. The method is based on the axis layout of parallel coordinates, with boxes representing the categories and parallelograms between the axes showing the relations between categories. In addition to the visual representation, we designed a rich set of interactions. Parallel Sets allow the user to interactively remap the data to new categorizations and, thus, to consider more data dimensions during exploration and analysis than usually possible. At the same time, a metalevel, semantic representation of the data is built. Common procedures, like building the cross product of two or more dimensions, can be performed automatically, thus complementing the interactive visualization. We demonstrate Parallel Sets by analyzing a large CRM data set, as well as investigating housing data from two US states. {\textcopyright} 2006 IEEE.},
 author = {Kosara, Robert and Bendix, Fabian and Hauser, Helwig},
