### 2020-01-17 #2017 @corybrunson
@clauswilke thank you for this feedback. I think i've addressed most of your concerns:

* In order to avoid undermining the change in later revisions, i'll do a find--replace from "i" to "I" once the paper is in its final form.
* I have corrected the outdated use of the `label.strata` parameter. Thank you for catching it!
* I have revised the "Concepts" section with a more careful and detailed comparison, in which i've tried to describe the ways ggalluvial is different from the other parallel sets packages in order to justify the classification scheme (Sankey, parallel sets, alluvial) i use but without suggesting that any is generally superior to the others. I think this addresses your third point; does it also address your fourth? I very much agree with you about the value of having multiple packages that accomplish similar but distinct things, but i do think one contrast in particular (allowing versus disallowing gaps and the meaningfulness of cumulative axis) is more significant than users tend to recognize, and which i think comes through most clearly in the third application, so i've made that more explicit.
