### 2020-01-17 b362e20
edit 'easy' footnote following edits to surrounding text
