+++ paper.md
@@ -82,7 +82,7 @@ The geoms of ggalluvial offer only a one-dimensional continuum between straight 
 The stats vertically arrange the elements (boxes and ribbons) at each axis. ggalluvial provides more levers of control over these statistical transformations, thereby over the messages conveyed by the plot, than the other packages.[^easy]
 Notably, in ggalluvial the plot elements are rendered by separate layers, following the additive (`+`) syntax of ggplot2, a feature shared by ggforce and ggpcp but not by ggparallel.
 
-[^easy]: Other ggplot2 extensions are not so flexible; in particular, the dependency package easyalluvial [@Koneswarakantha2019] was built on top of ggalluvial to exchange much of this flexibility for more expedient data exploration.
+[^easy]: Indeed, the dependency package easyalluvial [@Koneswarakantha2019] was built on top of ggalluvial to exchange much of this flexibility for more expedient data exploration.
 
 The ggalluvial package adopts the term _alluvial plot_ for the subtype of parallel sets plots it produces,[^plots] which are distinguished by two features: a strict order on the stacked elements at each axis, including both the values of the discrete variables and the ribbons connecting cases or cohorts between them; and a meaningful positional dimension along which these elements are stacked, which precludes inserting gaps between them. The first feature is shared by the other packages but are not essential to parallel sets plots; such plots could arrange boxes corresponding to repeated categorical decompositions differently at different axes. While most of the packages separate boxes at each axis with gaps, these can be reduced to zero, so that each package can create alluvial plots. (ggparallel and ggalluvial alone _only_ produce alluvial plots.)
 This feature is particularly important to some applications and, in my view, can fundamentally change the way a plot is interpreted.
