### 2020-01-17 #2017 @clauswilke
I am fine with the revision. One last comment: The "Concepts" section needs some copy-editing. In places it is difficult to follow the prose. For example, by the time I reach the sentence "... the plots generated by these extensions ..." I have completely lost track of what "these extensions" refers to.

You may benefit from talking through this section with a colleague, just so you get third-party feedback on whether the text flow can be followed or not.
