+++ paper.bib
@@ -242,6 +242,22 @@ title = {{alluvial: R Package for Creating Alluvial Diagrams}},
 url = {https://cran.r-project.org/package=alluvial},
 year = {2016}
 }
+@misc{Allaire2017,
+author = {Allaire, J. J. and Gandrud, Christopher and Russell, Kenton and Yetman, C. J.},
+title = {{networkD3: D3 JavaScript Network Graphs from R}},
+year = {2017}
+}
+@misc{Csardi2017,
+author = {Cs{\'{a}}rdi, G{\'{a}}bor and Weiner, January},
+title = {{sankey: Illustrate the Flow of Information or Material}},
+year = {2017}
+}
+@misc{Weiner2017,
+author = {Weiner, January},
+title = {{riverplot: Sankey or Ribbon Plots}},
+url = {https://cran.r-project.org/package=riverplot},
+year = {2017}
+}
 @article{Liu2018,
 author = {Liu, Ying and Zhao, Naizhuo and Vanos, Jennifer K and Cao, Guofeng},
 doi = {10.1177/0308518X17745067},
