+++ paper.rmd
@@ -110,37 +110,44 @@ Visualizations of flow processes have long encoded magnitudes as ribbon widths, 
 A widely-used subtype for longitudinal categorical data represent categories as nodes threaded by edges that represent the trajectories and magnitudes of cases [@Riehmann2005].
 Their design anticipated parallel sets plots, which were adapted from parallel coordinates plots [@Inselberg1987;@Wegman1990] to visualize multivariate categorical data, and which represent cohorts of equivalent cases as ribbons connecting categories represented as boxes [@Kosara2006].
 These in turn anticipated "alluvial diagrams", proposed to visualize changes in case memberships across successive cross-sections [@Rosvall2010].
-Several R packages have been developed to generate these diagrams, in particular alluvial [@Bojanowski2016], ggparallel [@Hofmann2013], ggforce [@Pedersen2019], ggalluvial [@Brunson2019], and ggpcp [@Ge2019].
+Several R packages have been developed to generate diagrams of these types, including riverplot [@Weiner2017], networkD3 [@Allaire2017], sankey [@Csardi2017], alluvial [@Bojanowski2016], ggparallel [@Hofmann2013], ggforce [@Pedersen2019], ggalluvial [@Brunson2019], and ggpcp [@Ge2019].
 
 Sankey, parallel sets, and alluvial diagrams are often conflated, and there is currently no consensus on what features are distinctive to each type.
 Moreover, their graphical elements go by a variety of names, often interchangeably.
-In order to more clearly describe the features of ggalluvial in relation to similar packages, it is useful to adopt a formal demarcation among these diagram types.
+In order to more clearly describe the features of ggalluvial in relation to similar packages, it is useful to adopt a careful demarcation among these diagram types.
 
 _Statistical graphics_ (here also simply called "plots") are diagrams that communicate statistical information using graphical methods [@Friendly2005] and, more narrowly, are uniquely determined from data by a fixed set of plotting rules [@Wilkinson2006].
-By design, graphics produced by ggplot2 extensions (and by the alluvial package) are plots: The stat, geom, and other layers of a ggplot object exactly reproduce a graphic from data (under the same parameter settings).[^exceptions]
-Sankey diagrams are much more flexible: The earliest engine efficiency diagrams in this tradition could take a variety of forms to depict the same energy flow and were differently annotated for different audiences [@Schmidt2008], and software implementations may be interactive and enable manual adjustments of graphical elements [@Riehmann2005].
-Moreover, the plots generated by these extensions are suitable for a wide variety of data types, whereas Sankey diagrams are still overwhelmingly used to represent flow. Arguably, the extensions are better understood as belonging to a different type of diagram.
+By design, graphics produced by ggplot2 extensions are plots: The stat, geom, and other layers of a ggplot object exactly reproduce a graphic from data (under the same parameter settings).[^exceptions]
+Sankey diagrams are much more flexible.
+The earliest engine efficiency diagrams in this tradition could take a variety of forms to depict the same energy flow and were differently annotated for different audiences [@Schmidt2008].
+Software implementations may use heuristic algorithms to position graphical elements [@Allaire2017;@Csardi2017] or enable users to manually, even interactively, adjust them [@Riehmann2005;@Allaire2017;@Weiner2017].
+Paradoxically, Sankey diagrams are overwhelmingly used to represent flow, whereas the aforecited ggplot2 extensions are used to visualize a wide variety of data types.
+Arguably, the extensions are better understood as producing a different type of diagram.
 
 [^exceptions]: There are exceptions. For example, the jitter geom in ggplot2 introduces randomness to symbol positions, and the repel geoms of ggrepel [@Slowikowski2019] use heuristic algorithms to position text.
 
[…]
