### 2020-01-18 #2017 @corybrunson
@clauswilke thank you again—i've made some more revisions that i'll run by a colleague and may revise further once the workweek resumes.
