### 2020-01-18 9fb1f7c
add urls to new software bibtex items
