+++ paper.bib
@@ -245,11 +245,13 @@ year = {2016}
 @misc{Allaire2017,
 author = {Allaire, J. J. and Gandrud, Christopher and Russell, Kenton and Yetman, C. J.},
 title = {{networkD3: D3 JavaScript Network Graphs from R}},
+url = {https://cran.r-project.org/package=networkD3},
 year = {2017}
 }
 @misc{Csardi2017,
 author = {Cs{\'{a}}rdi, G{\'{a}}bor and Weiner, January},
 title = {{sankey: Illustrate the Flow of Information or Material}},
+url = {https://cran.r-project.org/package=sankey},
 year = {2017}
 }
 @misc{Weiner2017,
