### 2020-01-20 5097c7c
small paper revisions
