+++ paper.md
@@ -79,19 +79,19 @@ Arguably, the extensions are better understood as producing a different type of 
 
 [^exceptions]: There are exceptions. For example, the jitter geom in ggplot2 introduces randomness to symbol positions, and the repel geoms of ggrepel [@Slowikowski2019] use heuristic algorithms to position text.
 
-Parallel sets plots might be viewed as a special subtype of Sankey diagram with the following features: Ribbons proceed monotonically along one dimension, and every ribbon encounters a box at every axis. These graphical constraints correspond to combinatorial constraints on the data, which amount to a id--key--value structure in which every id--key pair takes exactly one value (possibly 0 or `NA`).
+Parallel sets plots might be viewed as a subtype of Sankey diagram with the following features: Ribbons proceed monotonically along one dimension, and every ribbon encounters a box at every axis. These graphical constraints correspond to combinatorial constraints on the data, which amount to an id--key--value structure in which every id--key pair takes exactly one value (possibly zero or missing).
 In this sense, the plots produced by the ggplot2 extensions (and by the alluvial package) are parallel sets plots: Cohorts are partitioned into categories at each axis and connected by ribbons whose widths encode their magnitudes.[^hammock]
 
 The plots produced still vary---in the shapes of ribbons, the arrangements of boxes, and the presence of gaps between boxes at the same axis.
-The exceptional geoms of ggparallel offer common-angle as well as linear ribbons. Those of alluvial, ggforce, ggalluvial, and ggpcp offer a one-dimensional continuum between straight and x-spline ribbons.
+The exceptional geoms of ggparallel each offer common-angle as well as linear ribbons. Those of alluvial, ggforce, ggalluvial, and ggpcp offer a one-dimensional continuum between straight and x-spline ribbons.
 The stats vertically arrange the elements (boxes and ribbons) at each axis.
 These distinct elements are rendered by separate layers in ggforce, ggalluvial, and ggpcp, following the additive (`+`) syntax of ggplot2.
 ggalluvial provides more levers of control over these statistical transformations, thereby over the messages conveyed by the plot, than the other packages.[^easy]
 
 [^easy]: Indeed, the dependency package easyalluvial [@Koneswarakantha2019] was built on top of ggalluvial to exchange much of this flexibility for more expedient data exploration.
 
-The ggalluvial package adopts the term _alluvial plot_ for the subtype of parallel sets plots it produces,[^plots] with the geological terminology itemized above.
-These alluvial plots are distinguished by two features: a strict order on the stacked elements at each axis, including both the values of the discrete variables and the ribbons connecting cases or cohorts between them; and a meaningful positional dimension along which these elements are stacked, which precludes inserting gaps between them. The first feature is shared by the other packages but is not essential to parallel sets plots; such plots could arrange boxes corresponding to repeated categorical decompositions differently at different axes. While most of the packages separate boxes at each axis with gaps, these can be reduced to zero, so that each package can create alluvial plots. (ggparallel and ggalluvial alone _only_ produce alluvial plots.)
+The ggalluvial package adopts the term _alluvial plot_ for the subtype of parallel sets plots it produces,[^plots] with the geological terminology introduced above.
+These alluvial plots are distinguished by two features: a strict order on the stacked elements at each axis, including both the values of the discrete variables and the ribbons connecting cases or cohorts between them; and a real-valued plotting dimension perpendicular to that of flow, along which these elements are stacked, so that gaps between them are precluded. The first feature is shared by the other packages but is not essential to parallel sets plots; such plots could arrange boxes corresponding to repeated categorical decompositions differently at different axes. While most of the packages separate boxes at each axis with gaps, these can be reduced to zero, so that each package can create alluvial plots. (ggparallel and ggalluvial alone _only_ produce alluvial plots.)
[…]
