### 2020-01-26 4279abf
knot-fix param to control proportional vs fixed knot-pos - geom-alluvium + examples
