### 2020-01-27 6056fdb
extend knot-fix param to geom-flow + example needing debugging
