### 2020-01-29 43d8682
alternative curve options + debug knot-fix edits + document
