+++ R/geom-alluvium.r
@@ -21,6 +21,19 @@
 #' @param knot.fix Logical; whether to interpret `knot.pos` as a fixed value,
 #'   rather than (the default) as a proportion of the separation between
 #'   adjacent axes.
+#' @param curve Character; the type of curve used to produce flows. Defaults to
+#'   `"xspline"` and can be alternatively set to one of `"linear"`, `"cubic"`,
+#'   `"quintic"`, `"sine"`, `"arctangent"`, and `"sigmoid"`. Only the
+#'   `"xspline"` option uses the `knot.*` parameters, while only the alternative
+#'   curves use the `segments` parameter, and only `"arctangent"` and
+#'   `"sigmoid"` use the `reach` parameter.
+#' @param reach For alternative `curve`s based on asymptotic functions, the
+#'   value along the asymptote at which to truncate the function to obtain the
+#'   shape that will be scaled to fit between strata. Larger values result in
+#'   greater compression and steeper slopes. The `NULL` default will be changed
+#'   to `sqrt(3)` for `"arctangent"` and to `3` for `"sigmoid"`.
+#' @param segments The number of segments to be used in drawing each alternative
+#'   curve (each curved boundary of each flow).
 #' @example inst/examples/ex-geom-alluvium.r
 #' @export
 geom_alluvium <- function(mapping = NULL,
@@ -29,6 +42,7 @@ geom_alluvium <- function(mapping = NULL,
                           position = "identity",
                           width = 1/3,
                           knot.pos = 1/6, knot.fix = FALSE,
+                          curve = "xspline", reach = NULL, segments = 24,
                           na.rm = FALSE,
                           show.legend = NA,
                           inherit.aes = TRUE,
@@ -45,6 +59,9 @@ geom_alluvium <- function(mapping = NULL,
       width = width,
       knot.pos = knot.pos,
       knot.fix = knot.fix,
+      curve = curve,
+      reach = reach,
+      segments = segments,
       na.rm = na.rm,
       ...
     )
@@ -90,7 +107,8 @@ GeomAlluvium <- ggproto(
   
   draw_group = function(self, data, panel_scales, coord,
                         width = 1/3,
-                        knot.pos = 1/6, knot.fix = FALSE) {
+                        knot.pos = 1/6, knot.fix = FALSE,
+                        curve = "xspline", reach = NULL, segments = 24) {
     
     # add width to data
     data <- transform(data, width = width)
@@ -103,16 +121,19 @@ GeomAlluvium <- ggproto(
     
     if (nrow(data) == 1) {
       # spline coordinates (one axis)
-      spline_data <- with(data, data.frame(
+      curve_data <- with(data, data.frame(
         x = x + width / 2 * c(-1, 1, 1, -1),
         y = ymin + (ymax - ymin) * c(0, 0, 1, 1),
         shape = rep(0, 4)
       ))
-    } else {
+    } else if (curve %in% c("spline", "xspline")) {
       # spline coordinates (more than one axis)
-      spline_data <- data_to_xspl(data, knot.fix)
[…]
