+++ R/geom-flow.r
@@ -27,6 +27,7 @@ geom_flow <- function(mapping = NULL,
                       position = "identity",
                       width = 1/3,
                       knot.pos = 1/6, knot.fix = FALSE,
+                      curve = "xspline", reach = NULL, segments = 24,
                       aes.flow = "forward",
                       na.rm = FALSE,
                       show.legend = NA,
@@ -47,6 +48,9 @@ geom_flow <- function(mapping = NULL,
       width = width,
       knot.pos = knot.pos,
       knot.fix = knot.fix,
+      curve = curve,
+      reach = reach,
+      segments = segments,
       aes.flow = aes.flow,
       na.rm = na.rm,
       ...
@@ -84,11 +88,8 @@ GeomFlow <- ggproto(
   
   draw_panel = function(self, data, panel_params, coord,
                         width = 1/3, aes.flow = "forward",
-                        knot.pos = 1/6, knot.fix = FALSE) {
-    save(data, panel_params, coord,
-         width, aes.flow, knot.pos, knot.fix,
-         file = "draw-panel.rda")
-    load("draw-panel.rda")
+                        knot.pos = 1/6, knot.fix = FALSE,
+                        curve = "xspline", reach = NULL, segments = 24) {
     
     # exclude one-sided flows
     data <- data[complete.cases(data), ]
@@ -115,17 +116,19 @@ GeomFlow <- ggproto(
       function(x) factor(x, levels = unique(x))
     )), ]
     
-    # construct spline grobs
-    xspls <- plyr::alply(data, 1, function(row) {
+    # construct x-spline grobs
+    grobs <- plyr::alply(data, 1, function(row) {
       
-      # spline paths and aesthetics
-      xspl <- knots_to_xspl(row$xmax.0, row$xmin.1,
-                            row$ymin.0, row$ymax.0, row$ymin.1, row$ymax.1,
-                            row$knot.pos.0, row$knot.pos.1,
-                            knot.fix = knot.fix)
-      aes <- as.data.frame(row[flow_aes],
-                           stringsAsFactors = FALSE)[rep(1, 8), ]
-      f_data <- cbind(xspl, aes)
+      # path of spline or unit curve
+      f_path <- row_to_curve(row$xmax.0, row$xmin.1,
+                             row$ymin.0, row$ymax.0, row$ymin.1, row$ymax.1,
+                             row$knot.pos.0, row$knot.pos.1,
+                             curve = curve, reach = reach, segments = segments,
+                             knot.fix = knot.fix)
+      # aesthetics
+      aes <- as.data.frame(row[flow_aes], stringsAsFactors = FALSE)
+      # join aesthetics to path
+      f_data <- cbind(f_path, aes[rep(1, nrow(f_path)), ])
       
       # transform (after calculating spline paths)
       f_coords <- coord$transform(f_data, panel_params)
@@ -142,7 +145,7 @@ GeomFlow <- ggproto(
     })
     
     # combine spline grobs
-    grob <- do.call(grid::grobTree, xspls)
+    grob <- do.call(grid::grobTree, grobs)
     grob$name <- grid::grobName(grob, "xspline")
     grob
   },
@@ -150,18 +153,48 @@ GeomFlow <- ggproto(
   draw_key = draw_key_polygon
 )
 
[…]
