+++ man/stat_alluvium.Rd
@@ -288,13 +288,21 @@ gg +
 # irregular spacing between axes of a continuous variable
 data(Refugees, package = "alluvial")
 refugees_sub <- subset(Refugees, year \%in\% c(2003, 2005, 2010, 2013))
-ggplot(data = refugees_sub,
+gg <- ggplot(data = refugees_sub,
        aes(x = year, y = refugees, alluvium = country)) +
-  geom_alluvium(aes(fill = country),
-                alpha = .75, decreasing = FALSE, knot.pos = 1) +
-  geom_stratum(aes(stratum = country), decreasing = FALSE, width = 1/2) +
   theme_bw() +
   scale_fill_brewer(type = "qual", palette = "Set3")
+# proportional knot positioning (default)
+gg +
+  geom_alluvium(aes(fill = country),
+                alpha = .75, decreasing = FALSE) +
+  geom_stratum(aes(stratum = country), decreasing = FALSE, width = 1/2)
+# fixed knot positioning
+gg +
+  geom_alluvium(aes(fill = country),
+                alpha = .75, decreasing = FALSE,
+                knot.pos = 1, knot.fix = TRUE) +
+  geom_stratum(aes(stratum = country), decreasing = FALSE, width = 1/2)
 
 \dontrun{
 data(babynames, package = "babynames")
