+++ inst/examples/ex-geom-flow.r
@@ -7,11 +7,11 @@ ggplot(as.data.frame(Titanic),
   geom_stratum() + geom_text(stat = "stratum", infer.label = TRUE) +
   ggtitle("Alluvial plot of Titanic passenger demographic data")
 
-# use of facets
+# use of facets, with sigmoid flows
 ggplot(as.data.frame(Titanic),
        aes(y = Freq,
            axis1 = Class, axis2 = Sex)) +
-  geom_flow(aes(fill = Age), width = .4) +
+  geom_flow(aes(fill = Age), width = .4, curve = "sigmoid", reach = 6) +
   geom_stratum(width = .4) +
   geom_text(stat = "stratum", infer.label = TRUE, size = 3) +
   scale_x_discrete(limits = c("Class", "Sex")) +
@@ -30,12 +30,12 @@ ggplot(wph,
        aes(x = Year, alluvium = Region, y = Telephones)) +
   geom_flow(aes(fill = Region, colour = Region), width = 0)
 
-# rightward flow aesthetics for vaccine survey data
+# rightward flow aesthetics for vaccine survey data, with cubic flows
 data(vaccinations)
 levels(vaccinations$response) <- rev(levels(vaccinations$response))
 ggplot(vaccinations,
        aes(x = survey, stratum = response, alluvium = subject,
            y = freq, fill = response, label = round(a, 3))) +
-  geom_lode() + geom_flow() +
+  geom_lode() + geom_flow(curve = "cubic") +
   geom_stratum(alpha = 0) +
   geom_text(stat = "stratum")
