+++ tests/testthat/test-geom-flow.r
@@ -23,9 +23,9 @@ test_that("`geom_flow` draws correctly", {
   )
 })
 
+data(vaccinations)
+
 test_that("`geom_flow` orients flows correctly", {
-  data(vaccinations)
-  
   vdiffr::expect_doppelganger(
     "`geom_flow`: forward orientation",
     ggplot(vaccinations,
@@ -39,5 +39,50 @@ test_that("`geom_flow` orients flows correctly", {
            aes(x = survey, stratum = response, alluvium = subject,
                y = freq, fill = response)) +
       geom_lode() + geom_flow(aes.flow = "backward")
+  )
+})
+
+test_that("`geom_flow()` recognizes alternative curves", {
+  vdiffr::expect_doppelganger(
+    "`geom_flow`: 'identity' curve",
+    ggplot(vaccinations,
+           aes(x = survey, stratum = response, alluvium = subject,
+               y = freq, fill = response)) +
+      geom_stratum() + geom_flow(curve = "linear")
+  )
+  vdiffr::expect_doppelganger(
+    "`geom_flow`: 'identity' curve",
+    ggplot(vaccinations,
+           aes(x = survey, stratum = response, alluvium = subject,
+               y = freq, fill = response)) +
+      geom_stratum() + geom_flow(curve = "cubic")
+  )
+  vdiffr::expect_doppelganger(
+    "`geom_flow`: 'identity' curve",
+    ggplot(vaccinations,
+           aes(x = survey, stratum = response, alluvium = subject,
+               y = freq, fill = response)) +
+      geom_stratum() + geom_flow(curve = "quintic")
+  )
+  vdiffr::expect_doppelganger(
+    "`geom_flow`: 'identity' curve",
+    ggplot(vaccinations,
+           aes(x = survey, stratum = response, alluvium = subject,
+               y = freq, fill = response)) +
+      geom_stratum() + geom_flow(curve = "sine")
+  )
+  vdiffr::expect_doppelganger(
+    "`geom_flow`: 'identity' curve",
+    ggplot(vaccinations,
+           aes(x = survey, stratum = response, alluvium = subject,
+               y = freq, fill = response)) +
+      geom_stratum() + geom_flow(curve = "arctan")
+  )
+  vdiffr::expect_doppelganger(
+    "`geom_flow`: 'identity' curve",
+    ggplot(vaccinations,
+           aes(x = survey, stratum = response, alluvium = subject,
+               y = freq, fill = response)) +
+      geom_stratum() + geom_flow(curve = "sigmoid")
   )
 })
