+++ tests/testthat/test-geom-alluvium.r
@@ -30,3 +30,48 @@ test_that("`geom_alluvium` draws correctly", {
       geom_alluvium(aes(fill = country),  width = 0, knot.pos = 0)
   )
 })
+
+test_that("`geom_alluvium()` recognizes alternative curves", {
+  vdiffr::expect_doppelganger(
+    "`geom_alluvium`: 'identity' curve",
+    ggplot(vaccinations,
+           aes(x = survey, stratum = response, alluvium = subject,
+               y = freq, fill = response)) +
+      geom_alluvium(curve = "linear")
+  )
+  vdiffr::expect_doppelganger(
+    "`geom_alluvium`: 'identity' curve",
+    ggplot(vaccinations,
+           aes(x = survey, stratum = response, alluvium = subject,
+               y = freq, fill = response)) +
+      geom_alluvium(curve = "cubic")
+  )
+  vdiffr::expect_doppelganger(
+    "`geom_alluvium`: 'identity' curve",
+    ggplot(vaccinations,
+           aes(x = survey, stratum = response, alluvium = subject,
+               y = freq, fill = response)) +
+      geom_alluvium(curve = "quintic")
+  )
+  vdiffr::expect_doppelganger(
+    "`geom_alluvium`: 'identity' curve",
+    ggplot(vaccinations,
+           aes(x = survey, stratum = response, alluvium = subject,
+               y = freq, fill = response)) +
+      geom_alluvium(curve = "sine")
+  )
+  vdiffr::expect_doppelganger(
+    "`geom_alluvium`: 'identity' curve",
+    ggplot(vaccinations,
+           aes(x = survey, stratum = response, alluvium = subject,
+               y = freq, fill = response)) +
+      geom_alluvium(curve = "arctan")
+  )
+  vdiffr::expect_doppelganger(
+    "`geom_alluvium`: 'identity' curve",
+    ggplot(vaccinations,
+           aes(x = survey, stratum = response, alluvium = subject,
+               y = freq, fill = response)) +
+      geom_alluvium(curve = "sigmoid")
+  )
+})
