### 2020-01-30 6f3653c
document curves + tests
