### 2020-01-31 6553981
use interpolative splines with fewer segments for alternative curves
