### 2020-01-31 6d5b755
uniform shape param for spline approximations to alternative curves
