### 2020-02-02 4dc6127
correct test typos + revert to piecewise linear approximations to alternative curves
