+++ tests/testthat/test-geom-flow.r
@@ -44,42 +44,42 @@ test_that("`geom_flow` orients flows correctly", {
 
 test_that("`geom_flow()` recognizes alternative curves", {
   vdiffr::expect_doppelganger(
-    "`geom_flow`: 'identity' curve",
+    "`geom_flow`: 'linear' curve",
     ggplot(vaccinations,
            aes(x = survey, stratum = response, alluvium = subject,
                y = freq, fill = response)) +
       geom_stratum() + geom_flow(curve = "linear")
   )
   vdiffr::expect_doppelganger(
-    "`geom_flow`: 'identity' curve",
+    "`geom_flow`: 'cubic' curve",
     ggplot(vaccinations,
            aes(x = survey, stratum = response, alluvium = subject,
                y = freq, fill = response)) +
       geom_stratum() + geom_flow(curve = "cubic")
   )
   vdiffr::expect_doppelganger(
-    "`geom_flow`: 'identity' curve",
+    "`geom_flow`: 'quintic' curve",
     ggplot(vaccinations,
            aes(x = survey, stratum = response, alluvium = subject,
                y = freq, fill = response)) +
       geom_stratum() + geom_flow(curve = "quintic")
   )
   vdiffr::expect_doppelganger(
-    "`geom_flow`: 'identity' curve",
+    "`geom_flow`: 'sine' curve",
     ggplot(vaccinations,
            aes(x = survey, stratum = response, alluvium = subject,
                y = freq, fill = response)) +
       geom_stratum() + geom_flow(curve = "sine")
   )
   vdiffr::expect_doppelganger(
-    "`geom_flow`: 'identity' curve",
+    "`geom_flow`: 'arctangent' curve",
     ggplot(vaccinations,
            aes(x = survey, stratum = response, alluvium = subject,
                y = freq, fill = response)) +
       geom_stratum() + geom_flow(curve = "arctan")
   )
   vdiffr::expect_doppelganger(
-    "`geom_flow`: 'identity' curve",
+    "`geom_flow`: 'sigmoid' curve",
     ggplot(vaccinations,
            aes(x = survey, stratum = response, alluvium = subject,
                y = freq, fill = response)) +
