### 2020-02-03 038c83b
change knot-pos default to 1/4 + manage vdiff cases
