+++ R/geom-flow.r
@@ -26,7 +26,7 @@ geom_flow <- function(mapping = NULL,
                       stat = "flow",
                       position = "identity",
                       width = 1/3,
-                      knot.pos = 1/6, knot.fix = FALSE,
+                      knot.pos = 1/4, knot.fix = FALSE,
                       curve = "xspline", reach = NULL, segments = 48,
                       aes.flow = "forward",
                       na.rm = FALSE,
@@ -88,7 +88,7 @@ GeomFlow <- ggproto(
   
   draw_panel = function(self, data, panel_params, coord,
                         width = 1/3, aes.flow = "forward",
-                        knot.pos = 1/6, knot.fix = FALSE,
+                        knot.pos = 1/4, knot.fix = FALSE,
                         curve = "xspline", reach = NULL, segments = 48) {
     
     # exclude one-sided flows
