### 2020-02-04 3d0a179
increase default reach + debug aes-bind examples + document
