### 2020-02-05 440cf04
paper - declutter refs + add curves footnote
