+++ paper.bib
@@ -29,13 +29,11 @@ url = {http://www.tandfonline.com/doi/abs/10.1080/01621459.1990.10474926},
 volume = {85},
 year = {1990}
 }
-@misc{Friendly2002,
-abstract = {This article provides an illustrated history of the visual and conceptual ideas leading to the development of mosaic displays. We trace the origins of the use of rectangles and area to depict data quantities and their relations, of early forms of mosaic displays including subdivided bar-like charts and various cartograms, to the modern forms used in log-linear analysis and in space-filling tree maps.},
+@article{Friendly2002,
 author = {Friendly, Michael},
-booktitle = {Journal of Computational and Graphical Statistics},
+journal = {Journal of Computational and Graphical Statistics},
 doi = {10.1198/106186002317375631},
 issn = {10618600},
-keywords = {Cartogram,Data visualization,Log-linear models,Mosaic matrix,Space-filling displays,Thematic cartography,Tree map},
 number = {1},
 pages = {89--107},
 title = {{A Brief History of the Mosaic Display}},
@@ -44,7 +42,6 @@ volume = {11},
 year = {2002}
 }
 @inproceedings{Riehmann2005,
-abstract = {We present a system that allows users to interactively explore complex flow scenarios represented as Sankey diagrams. Our system provides an overview of the flow graph and allows users to zoom in and explore details on demand. The support for quantitative flow tracing across the flow graph as well as representations at different levels of detail facilitate the understanding of complex flow situations. The energy flow in a city serves as a sample scenario for our system. Different forms of energy are distributed within the city and they are transformed into heat, electricity, or other forms of energy. These processes are visualized and interactively explored. In addition our system can be used as a planning tool for the exploration of alternative scenarios by interactively manipulating different parameters in the energy flow network.},
 author = {Riehmann, P. and Hanfler, M. and Froehlich, B.},
 booktitle = {IEEE Symposium on Information Visualization, 2005. INFOVIS 2005.},
 doi = {10.1109/INFVIS.2005.1532152},
@@ -56,7 +53,6 @@ url = {http://ieeexplore.ieee.org/document/1532152/},
 year = {2005}
 }
 @incollection{Friendly2005,
-abstract = {The Milestones Project is a comprehensive attempt to collect, document, illustrate, and interpret the historical developments leading to modern data visualization and visual thinking. This paper provides an overview and brief tour of the milestones content, with a few illustrations of significant contributions to the history of data visualization. This forms one basis for exploring interesting questions and problems in the use of statistical and graphical methods to explore this history, a topic that can be called “statistical historiography.”},
 address = {Berlin/Heidelberg},
 author = {Friendly, Michael},
 booktitle = {Classification — the Ubiquitous Challenge. Studies in Classification, Data Analysis, and Knowledge Organization},
@@ -72,7 +68,6 @@ author = {Wilkinson, Leland},
 edition = {2},
 doi = {10.1007/0-387-28695-0},
 isbn = {0387286950},
-pages = {691},
 publisher = {Springer Science {\&} Business Media},
 title = {{The Grammar of Graphics}},
 url = {https://books.google.com/books?id=NRyGnjeNKJIC},
@@ -84,7 +79,6 @@ title = {{ggrepel: Automatically Position Non-Overlapping Text Labels with 'ggpl
 year = {2019}
 }
 @inproceedings{Kosara2006,
[…]
