### 2020-02-10 #2017 @corybrunson
@danielskatz thanks for the prompt. I believe i've addressed the first reviewer's concerns, and i should be able to address any from the second reviewer within a few days' time. (I'm traveling but will only be a few days at a time without wifi.)

Incidentally, i've almost finished an upgrade on a new branch that i didn't think i'd be able to get to for some time (alternative curves for plotting flows; i mention this in the paper as the only planned feature not yet implemented). If the review is still open, would it be OK for me to merge this branch in? It would increment the version number.
