### 2020-02-16 bd16d63
default segments to null since unused when curve defaults to xspline
