### 2020-02-16 b253092
test interpolate aesthetic flow option
