+++ R/geom-flow.r
@@ -28,13 +28,14 @@ geom_flow <- function(mapping = NULL,
                       width = 1/3,
                       knot.pos = 1/4, knot.fix = FALSE,
                       curve = "xspline", reach = NULL, segments = NULL,
+                      space = "rgb", interpolate = "linear",
                       aes.flow = "forward",
                       na.rm = FALSE,
                       show.legend = NA,
                       inherit.aes = TRUE,
                       ...) {
   
-  aes.flow <- match.arg(aes.flow, c("forward", "backward"))
+  aes.flow <- match.arg(aes.flow, c("forward", "backward", "interpolate"))
   
   layer(
     geom = GeomFlow,
@@ -51,6 +52,8 @@ geom_flow <- function(mapping = NULL,
       curve = curve,
       reach = reach,
       segments = segments,
+      space = space,
+      interpolate = interpolate,
       aes.flow = aes.flow,
       na.rm = na.rm,
       ...
@@ -89,27 +92,50 @@ GeomFlow <- ggproto(
   draw_panel = function(self, data, panel_params, coord,
                         width = 1/3, aes.flow = "forward",
                         knot.pos = 1/4, knot.fix = FALSE,
-                        curve = "xspline", reach = NULL, segments = NULL) {
+                        curve = "xspline", reach = NULL, segments = NULL,
+                        space = "rgb", interpolate = "linear") {
+    save(data, panel_params, coord,
+         width, aes.flow, knot.pos, knot.fix, curve, reach, segments,
+         space, interpolate,
+         file = "draw-panel.rda")
+    load("draw-panel.rda")
     
     # exclude one-sided flows
     data <- data[complete.cases(data), ]
+    
+    # ensure compatibility among parameters
+    if (aes.flow == "interpolate" && curve == "xspline") {
+      stop("Cannot interpolate flow aesthetics along an x-spline curve; ",
+           "use an alternative curve instead.")
+    }
+    if (curve != "xspline") {
+      # default to 48 segments per curve, ensure the minimum number of segments
+      if (is.null(segments)) segments <- 48 else if (segments < 3) {
+        #warning("Must use at least 3 segments; substituting `segments = 3`.")
+        segments <- 3
+      }
+    }
     
     # adjoin data with itself by alluvia along adjacent axes
     flow_pos <- intersect(names(data), c("x", "xmin", "xmax", "width",
                                          "y", "ymin", "ymax", "knot.pos"))
     flow_aes <- intersect(names(data), c("size", "linetype",
                                          "colour", "fill", "alpha"))
-    flow_fore <- if (aes.flow != "backward") flow_aes else NULL
-    flow_back <- if (aes.flow != "forward") flow_aes else NULL
+    flow_fore <- if (aes.flow == "forward") flow_aes else NULL
+    flow_back <- if (aes.flow == "backward") flow_aes else NULL
+    flow_link <- c(flow_pos, if (aes.flow == "interpolate") flow_aes)
     data <- self_adjoin(
       data = data, key = "x", by = "alluvium",
-      link = flow_pos,
+      link = flow_link,
       keep.x = flow_fore, keep.y = flow_back,
       suffix = c(".0", ".1")
     )
     
     # aesthetics (in prescribed order)
-    aesthetics <- intersect(.color_diff_aesthetics, names(data))
+    aesthetics <- intersect(
+      if (aes.flow == "interpolate") paste0(flow_aes, ".0") else flow_aes,
+      names(data)
+    )
     # arrange data by aesthetics for consistent (reverse) z-ordering
     data <- data[do.call(order, lapply(
[…]
