+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: "Layered Grammar for Alluvial Plots"
+title: "ggalluvial: Layered Grammar for Alluvial Plots"
 tags:
   - R
   - ggplot2
@@ -34,7 +34,7 @@ Alluvial diagrams use stacked bar plots and variable-width ribbons to represent 
 
 The package makes two key contributions to the R ecosystem.
 First, ggalluvial anchors the imprecise notion of an alluvial diagram to the rigid grammar of graphics [@Wilkinson2006], which lends the plots more precise meaning and opens up many combinatorial possibilities.
-Second, ggalluvial adopts a distinctive geological nomenclature to distinguish "alluvial plots" and their graphical elements from Sankey diagrams and parallel sets plots, which i hope prove useful as these visualization tools converge toward common standards.
+Second, ggalluvial adopts a distinctive geological nomenclature to distinguish "alluvial plots" and their graphical elements from Sankey diagrams and parallel sets plots, which I hope prove useful as these visualization tools converge toward common standards.
 
 ## Functionality
 
@@ -94,9 +94,9 @@ ggalluvial provides more levers of control over these statistical transformation
 The ggalluvial package adopts the term _alluvial plot_ for the subtype of parallel sets plots it produces,[^plots] with the geological terminology introduced above.
 These alluvial plots are distinguished by two features: a strict order on the stacked elements at each axis, including both the values of the discrete variables and the ribbons connecting cases or cohorts between them; and a real-valued plotting dimension perpendicular to that of flow, along which these elements are stacked, so that gaps between them are precluded. The first feature is shared by the other packages but is not essential to parallel sets plots; such plots could arrange boxes corresponding to repeated categorical decompositions differently at different axes. While most of the packages separate boxes at each axis with gaps, these can be reduced to zero, so that each package can create alluvial plots. (ggparallel and ggalluvial alone _only_ produce alluvial plots.)
 This feature is particularly important to some applications and, in my view, can fundamentally change the way a plot is interpreted.
-It is for this reason that i believe the new typology and terminology are warranted.
+It is for this reason that I believe the new typology and terminology are warranted.
 
-[^hammock]: The possible exceptions are the hammock plots and common angle plots of ggparallel, which are contrasted with a stricter definition of parallel sets plots than i use here, in which ribbons are straight, their widths aggregate to box widths, and they meet without overlap at the sides of boxes, partitioning them [@Hofmann2013].
+[^hammock]: The possible exceptions are the hammock plots and common angle plots of ggparallel, which are contrasted with a stricter definition of parallel sets plots than I use here, in which ribbons are straight, their widths aggregate to box widths, and they meet without overlap at the sides of boxes, partitioning them [@Hofmann2013].
 [^plots]: This has the unfortunate side effect of conflating search results from the geology literature.
 
 ## Applications
@@ -116,10 +116,10 @@ Alluvial plots clearly indicate times at which longitudinal data are censored or
 **Signed categorical data:**
[…]
