### 2020-02-19 6be922d
debug irregular x in stat-flow + add tests
