### 2020-03-03 corybrunson/ggalluvial#48 @tweed1e
Add community guidelines for contributing/reporting
The JOSS code review requires (https://github.com/openjournals/joss-reviews/issues/2017):

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I know you mention these in the README.md, but there could also be a specific CONTRIBUTING.md file.

I'm not sure what the best practice is (it seems like github has features that suggest its own community guidelines, etc.) but here is an example I found for R from ropensci+JOSS: https://github.com/ropensci/dotgithubfiles

There may be more recent examples or better practices as well.
