### 2020-03-03 corybrunson/ggalluvial#48 @corybrunson
I've vacillated on this and it's probably good to be prompted to go ahead with it. Since this is a ggplot2 extension, i'll adapt a `CONTRIBUTING.md` file from theirs ([here](https://github.com/tidyverse/ggplot2/blob/master/CONTRIBUTING.md)), though since it'll be shorter i'm glad to have the ropensci one as a point of reference. (I also haven't come across any widely-used best practice.) Thanks! I'll get to this today or tomorrow.
