### 2020-03-03 #2017 @tweed1e
I second the copy-editing request of the other reviewer, and my last hiccup:

I get really confused by the terminology. I know about ggplot2 and its extensions, I know a bit about sankey diagrams and similar things, so I think I'm kind of the non-specialist audience that would need to understand the paper / documentation. I get the need for specific terminology to nail down exactly what the terms mean so people can understand the differences between a parallel sets, alluvial, and sankeys, but in combo with the prose (see @clauswilke's comment above), it's hard for the general audience (non-geologists, I guess?) to grasp what the purpose is and what the individual parts are without doing a deep dive into viz generally and the package itself.

I'll make a few specific requests that I think will help:

- Make an alluvial diagram (maybe just re-use the Titanic one) with labels pointing to an alluvium, a lode, a flow, a strata. I thought that's what the diagram on page 2 was going to be, so I got extra confused when a stratum has a 'flow' label, etc etc.
- For the plot on page 2: add a label (e.g., Figure 1: Alluvial plot relating geoms and stats) and reference that label in the text (e.g., "The alluvial plot in Figure 1" instead of "the following alluvial plot")

Then I'll check off "Summary" and "Quality of writing" and accept ✅👍.
