### 2020-03-03 #2017 @clauswilke
Since I also haven't checked "quality of writing": I'm mostly waiting to look at the final version. I think the comments by @tweed1e are helpful, and I expect that I can check off "quality of writing" once they have been addressed.
