+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,11 @@
+# Contributor Code of Conduct
+
+As contributors and maintainers of this project, we pledge to respect all people who contribute through opening issues (including questions, bugs, and feature requests), updating documentation, submitting pull requests or patches, and other activities.
+
+We are committed to making participation in this project a harassment-free experience for everyone, regardless of level of experience or fluency, contribution history, social status, age, gender identity or expression, sexual orientation, visible or invisible disability, personal appearance, body size, ethnic or racial identity or perception, and cultural or religious background or identity.
+
+Examples of unacceptable behavior by participants include the use of sexual language or imagery, derogatory or derisive comments, personal attacks, trolling, public or private harassment or insults, or other unprofessional conduct.
+
+Project maintainers have the right and responsibility to remove, edit, or reject comments, commits, code, issues, and other contributions inconsistent with this Code of Conduct. Project maintainers who do not follow this Code of Conduct may be removed from the project team. Instances of behavior inconsistent with this Code of Conduct may be reported by opening an issue or contacting one or more of the project maintainers.
+
+This Code of Conduct is adapted (with changes) from the [Contributor Covenant](http:contributor-covenant.org), version 1.0.0, available at https://www.contributor-covenant.org/version/1/0/0/code-of-conduct.html
