### 2020-03-03 corybrunson/ggalluvial#48 @corybrunson
The most recent commit to `master` (01aead6776440b27664d5cd20a1f16663fa32c4e) includes `CONTRIBUTING.md` and `CODE_OF_CONDUCT.md`, both loosely adapted from the sources mentioned here. Please let me know if you think they could be improved!
