### 2020-03-03 #2017 @corybrunson
@tweed1e thank you! To address your first request, do you think it would suffice to lightly edit the "Alluvial plots" section of [the primary vignette](http://corybrunson.github.io/ggalluvial/articles/ggalluvial.html)? I originally erred toward concision but i certainly don't mind adding useful content to the paper. If that makes sense, then i'll also add figure labels, then push the revised paper and have whedon re-generate the PDF.
