### 2020-03-10 df625ab
rm plyr dependency - addresses #49
