+++ R/geom-lode.r
@@ -76,7 +76,7 @@ GeomLode <- ggproto(
     )
     
     # construct polygon grobs
-    polys <- plyr::alply(data, 1, function(row) {
+    polys <- lapply(split(data, seq_len(nrow(data))), function(row) {
       
       poly <- rect_to_poly(row$xmin, row$xmax, row$ymin, row$ymax)
       aes <- as.data.frame(row[lode_aes],
