### 2020-03-16 49c4392
more illustrative example in paper
