+++ paper.rmd
@@ -51,53 +51,31 @@ Second, ggalluvial adopts a distinctive geological nomenclature to distinguish "
 * _lodes_, subdivisions of strata by their intersections with alluvia
 * _flows_, segments of alluvia between strata
 
-Plot layers are formed by pairing stats (statistical transformations) with geoms (mappings to graphical elements and properties); while every stat and geom has a conventional default, alternative grammatical pairings provide combinatorial richness to plotting possibilities.
-The following alluvial plot depicts several meaningful stat--geom combinations from the documentation. Default pairings, other within-package ("alluvial") pairings, and pairings of alluvial stats with other geoms are differentiated by color:
+These layers, and other plot elements, are illustrated in the alluvial plot below, which visualizes changes in several students' curricula (based on their declared majors) across several academic terms:
 
-```{r figure, echo=FALSE, message=FALSE, warning=FALSE, fig.height=4, fig.width=8, fig.cap=""}
+```{r figure, echo=FALSE, message=FALSE, warning=FALSE, fig.height=4.5, fig.width=8, fig.cap=""}
 library(ggalluvial)
-combos <- data.frame(
-  stat = c(
-    rep("stratum", 2),
-    rep("alluvium", 4),
-    "flow",
-    # text geom
-    c("stratum", "alluvium", "flow"),
-    # errorbar and pointrange geoms
-    "stratum", "alluvium", "flow"
-  ),
-  geom = c(
-    c("stratum", "lode"),
-    c("stratum", "alluvium", "lode", "flow"),
-    "flow",
-    rep("text", 3),
-    "errorbar", "pointrange", "pointrange"
-  ),
-  pairing = c(
-    c("default", "alluvial"),
-    c("alluvial", "default", "default", "alluvial"),
-    "default",
-    rep("external", 6)
-  )
-)
-levs <- c("stratum", "alluvium", "lode", "flow",
-          "text", "errorbar", "pointrange")
-combos <- transform(
-  combos,
-  stat = factor(stat, levs),
-  geom = factor(geom, levs),
-  pairing = factor(pairing, c("default", "alluvial", "external"))
-)
-ggplot(combos, aes(axis1 = stat, axis2 = geom)) +
-  scale_x_discrete(limits = c("stat", "geom"), expand = c(.1, .05)) +
-  geom_alluvium(aes(fill = pairing)) +
-  scale_fill_brewer(type = "qual") +
-  geom_stratum() +
-  geom_text(stat = "stratum", infer.label = TRUE) +
-  theme_minimal()
+data(majors)
+majors$curriculum <- as.factor(majors$curriculum)
+ggplot(majors,
+       aes(x = semester, stratum = curriculum, alluvium = student,
+           fill = curriculum)) +
+  scale_fill_brewer(type = "qual", palette = "Set2") +
+  scale_y_continuous(breaks = seq(-2, 10, 2)) +
+  geom_flow(stat = "alluvium", lode.guidance = "frontback",
+            color = "darkgray", negate.strata = NA) +
+  geom_stratum(negate.strata = NA) +
+  theme(legend.position = "right") +
+  labs(x = NULL) +
+  ggtitle("Changes in curricula in a student cohort")
 ```
 
-Alluvial layers can interpret two forms of tidy data, long (one row per lode) and wide (one row per alluvium), related by the pivot operations of tidyr [@Wickham2019].
[…]
