+++ paper.md
@@ -45,11 +45,10 @@ Second, ggalluvial adopts a distinctive geological nomenclature to distinguish "
 * _lodes_, subdivisions of strata by their intersections with alluvia
 * _flows_, segments of alluvia between strata
 
-These layers, and other plot elements, are illustrated in the alluvial plot below, which visualizes changes in several students' curricula (based on their declared majors) across several academic terms:
-
-![](paper/figure-1.png)
-
+Figure \ref{fig:example} illustrates these and other plot elements by visualizing changes in several students' curricula (based on their declared majors) across several academic terms.
 Each axis corresponds to an odd-valued term (1 through 15), at which the students are grouped into strata according to their curricula---Art History, Ceramic, etc. The individual students can be tracked from term to term along their alluvia: for instance, one student started out in Digital Art, encoded by the blue ribbon, but switched to Painting in the 11th term, where the ribbon turns pink. The partially transparent flows are colored according to their originating (not their terminating) terms, and the lodes where they intersect the strata are obscured by the solid-colored strata themselves. When a student's curriculum is unknown, they are grouped into the "missing" (`NA`) stratum, which is weighted negatively in this example.
+
+![\label{fig:example}Alluvial plot of changes in curricula in a student cohort](paper/figure-1.png)
 
 Plot layers are formed by pairing stats (statistical transformations) with geoms (mappings to graphical elements and properties); while every stat and geom has a conventional default, alternative grammatical pairings provide combinatorial richness to plotting possibilities. In the above example, the alluvium geom was paired with the flow stat, so that the flows of each alluvium could change color across the axes.
 Other meaningful stat--geom combinations can be found in the documentation, including pairings of the three alluvial stats (stratum, alluvium, and flow) with the text, errorbar, and pointrange geoms.
