### 2020-03-16 #2017 @corybrunson
@clauswilke thank you! @tweed1e i've addressed some of your concerns with the new example, adapted from a data set that comes with the package. I did away with the alluvial plot of geom–stat pairings, since it was contrived in the first place to provide a plot for the paper. I can add labels and pointers if it would add to the quick-readability, but i wanted to first push a version that's reproducible from code.
