### 2020-03-19 23a291e
compute count & prop for after-stat
