### 2020-03-19 b1741ba
compute n for after-stat + update documentation
