+++ man-roxygen/computed-variables.r
@@ -1,0 +1,9 @@
+#' @section Computed variables:
+#' These can be used with [ggplot2::after_stat()] to [control aesthetic
+#' evaluation](https://ggplot2.tidyverse.org/reference/aes_eval.html).
+#' \describe{
+#'   \item{n}{number of cases in lode}
+#'   \item{count}{cumulative weight of lode}
+#'   \item{prop}{weighted proportion of lode}
+#' }
+#' Each variable is grouped by axis (the value of `x`).
