### 2020-03-19 18f4091
add weight param to stat layers, remove from data before output - addresses #50
