### 2020-03-19 a446bc7
add tests for computed variables
