### 2020-03-19 1d514b3
update news re computed variables
