### 2020-03-19 2728d62
simplify tests for computed variables
