+++ tests/testthat/test-stat-stratum.r
@@ -3,17 +3,14 @@ context("stat-stratum")
 # weights are used but not returned
 
 test_that("`stat_stratum` weights computed variables but drops weight", {
-  data <- data.frame(x = rep(1:2, c(2, 3)),
-                     stratum = LETTERS[c(1, 2, 1, 2, 2)])
-  data$y <- c(1, 2, 1, 2, 3)
-  data$weight <- c(.5, 1, 2.5, 2, 1.5)
+  data <- data.frame(x = rep(1:2, c(2, 3)), stratum = LETTERS[c(1, 2, 1, 2, 2)])
+  data$y <- c(1, 1, 1, 1, 2)
+  data$weight <- c(.5, 1, .5, 1, 1.5)
   comp <- StatStratum$compute_panel(data)
-  expect_equivalent(comp[with(comp, order(x, stratum)), ]$n,
-                    c(0.5, 1, 2.5, 3.5))
-  expect_equivalent(comp[with(comp, order(x, stratum)), ]$count,
-                    c(0.5, 2, 2.5, 8.5))
-  expect_equivalent(comp[with(comp, order(x, stratum)), ]$prop,
-                    c(2.2, 8.8, 2.5, 8.5) / 11)
+  comp <- comp[with(comp, order(x, stratum)), ]
+  expect_equivalent(comp$n, c(0.5, 1, 0.5, 2.5))
+  expect_equivalent(comp$count, c(0.5, 1, 0.5, 4))
+  expect_equivalent(comp$prop, c(c(1, 2) / 3, c(1, 8) / 9))
   expect_null(comp$weight)
 })
 
