+++ tests/testthat/test-stat-alluvium.r
@@ -6,25 +6,21 @@ test_that("`stat_alluvium` weights computed variables but drops weight", {
   # not cementing alluvia
   data <- expand.grid(alluvium = letters[1:3], x = 1:2)
   data$stratum <- LETTERS[c(1, 1, 2, 1, 2, 2)]
-  data$y <- c(1, 2, 1, 3, 1, 2)
-  data$weight <- c(.5, 1, 2.5, 2, 1.5, 1)
+  data$y <- c(1, 1, 1, 1, 1, 2)
+  data$weight <- c(.5, 1, 1, .5, 1, 1)
   comp <- StatAlluvium$compute_panel(data)
-  expect_equivalent(comp[with(comp, order(x, alluvium)), ]$n,
-                    c(0.5, 1, 2.5, 2, 1.5, 1))
-  expect_equivalent(comp[with(comp, order(x, alluvium)), ]$count,
-                    c(0.5, 2, 2.5, 6, 1.5, 2))
-  expect_equivalent(comp[with(comp, order(x, alluvium)), ]$prop,
-                    c(1.9, 7.6, 9.5, 12, 3, 4) / 19)
+  comp <- comp[with(comp, order(x, alluvium)), ]
+  expect_equivalent(comp$n, c(0.5, 1, 1, 0.5, 1, 1))
+  expect_equivalent(comp$count, c(0.5, 1, 1, 0.5, 1, 2))
+  expect_equivalent(comp$prop, c(c(1, 2, 2) / 5, c(1, 2, 4) / 7))
   expect_null(comp$weight)
   # cementing alluvia
   data$stratum <- LETTERS[c(1, 1, 2)]
   comp <- StatAlluvium$compute_panel(data, cement.alluvia = TRUE)
-  expect_equivalent(comp[with(comp, order(x, alluvium)), ]$n,
-                    c(1.5, 2.5, 3.5, 1))
-  expect_equivalent(comp[with(comp, order(x, alluvium)), ]$count,
-                    c(2.5, 2.5, 7.5, 2))
-  expect_equivalent(comp[with(comp, order(x, alluvium)), ]$prop,
-                    c(9.5, 9.5, 15, 4) / 19)
+  comp <- comp[with(comp, order(x, alluvium)), ]
+  expect_equivalent(comp$n, c(1.5, 1, 1.5, 1))
+  expect_equivalent(comp$count, c(1.5, 1, 1.5, 2))
+  expect_equivalent(comp$prop, c(c(3, 2) / 5, c(3, 4) / 7))
   expect_null(comp$weight)
 })
 
