### 2020-03-20 0a23a80
add calculated variable lode for labeling purposes + asoc. distill param
