+++ R/utils.r
@@ -3,3 +3,10 @@
   "fill", "bg", "fg", "col", "colour", "color", "alpha",
   "lty", "linetype", "cex", "lwd", "size", "pch", "shape"
 )
+
+# text aesthetics
+.text_aesthetics <- c(
+  "label",
+  "vjust", "hjust", "angle",
+  "family", "fontface", "lineheight"
+)
