### 2020-03-20 488be3f
deprecate infer-label param + defunct label-strata param
