+++ inst/examples/ex-alluvial-data.r
@@ -48,11 +48,11 @@ gg <- ggplot(majors_alluvia,
 gg +
   geom_alluvium(aes(fill = as.factor(student)), width = 2/5, discern = TRUE) +
   geom_stratum(width = 2/5, discern = TRUE) +
-  geom_text(stat = "stratum", discern = TRUE, infer.label = TRUE)
+  geom_text(stat = "stratum", discern = TRUE, aes(label = after_stat(stratum)))
 gg +
   geom_alluvium(aes(fill = as.factor(student)), width = 2/5, discern = FALSE) +
   geom_stratum(width = 2/5, discern = FALSE) +
-  geom_text(stat = "stratum", discern = FALSE, infer.label = TRUE)
+  geom_text(stat = "stratum", discern = FALSE, aes(label = after_stat(stratum)))
 # warning when inappropriate
 ggplot(majors[majors$semester %in% paste0("CURR", c(1, 7, 13)), ],
        aes(x = semester, stratum = curriculum, alluvium = student,
