+++ man-roxygen/computed-variables.r
@@ -5,6 +5,8 @@
 #'   \item{n}{number of cases in lode}
 #'   \item{count}{cumulative weight of lode}
 #'   \item{prop}{weighted proportion of lode}
+#'   \item{stratum}{stratum value}
+#'   \item{alluvium}{alluvium value}
 #'   \item{lode}{lode label distilled from alluvia}
 #' }
 #' The numerical variables are grouped by `x` and weighted by `weight`. The
