### 2020-03-22 be51af1
annotate figure in joss paper
