### 2020-03-22 #2017 @corybrunson
@tweed1e i learned how to use `ggplot2::annotate()` and added labels and arrows to the plot! I think this is what you had in mind.
