### 2020-03-22 #2017 @tweed1e
Yep, that looks better. Still a bit confused but the labels are helping---e.g., it seems the alluvia and flows are the same until I notice the label "*one* alluvium", and that helps me tell the difference.

Last request: remove the "y" axis label, and make the x-axis labels consistent with the text---the text says "Term", but the x-axis tick labels say "CURR1", etc. So either change "CURR1" to "Term 1", or add "Term" as the x-axis label, and 1, 3, 5, as the x-axis tick labels. Otherwise it should be good to go! 👍
