### 2020-03-22 60e0fc5
rm alluvium as a computed variable
