+++ R/stat-alluvium.r
@@ -175,7 +175,7 @@ StatAlluvium <- ggproto(
     # introduce label
     if (infer.label) {
       deprecate_parameter("infer.label",
-                          msg = "Use `aes(label = after_stat(alluvium))`.")
+                          msg = "Use `aes(label = after_stat(lode))`.")
       if (is.null(data$label)) {
         data$label <- data$alluvium
       } else {
