+++ R/stat-flow.r
@@ -145,7 +145,7 @@ StatFlow <- ggproto(
     # introduce label
     if (infer.label) {
       deprecate_parameter("infer.label",
-                          msg = "Use `aes(label = after_stat(alluvium))`.")
+                          msg = "Use `aes(label = after_stat(lode))`.")
       if (is.null(data$label)) {
         data$label <- data$alluvium
       } else {
