### 2020-03-22 1923612
clarify labels in joss paper figure
