+++ paper.rmd
@@ -54,21 +54,22 @@ Second, ggalluvial adopts a distinctive geological nomenclature to distinguish "
 Figure \ref{fig:example} illustrates these and other plot elements by visualizing changes in several students' curricula (based on their declared majors) across several academic terms.
 Each axis corresponds to an odd-valued term (1 through 15), at which the students are grouped into strata according to their curricula---Art History, Ceramic, etc. The individual students can be tracked from term to term along their alluvia: for instance, one student started out in Digital Art, encoded by the blue ribbon, but had switched to Painting by the 11th term, where the ribbon turns pink. The partially transparent flows are colored according to their originating (not their terminating) terms, and the lodes where they intersect the strata are obscured by the solid-colored strata themselves. When a student's curriculum is unknown, they are grouped into the "missing" (`NA`) stratum, which is weighted negatively in this example.
 
-```{r figure, echo=FALSE, message=FALSE, warning=FALSE, fig.height=4, fig.width=8, fig.cap="\\label{fig:example}Alluvial plot of changes in curricula in a student cohort"}
+```{r figure, echo=FALSE, message=FALSE, warning=FALSE, fig.height=4, fig.width=8, fig.cap="\\label{fig:example}Alluvial plot of changes in curricula by a cohort of art students"}
 library(ggalluvial)
 data(majors)
 majors$curriculum <- as.factor(majors$curriculum)
+levels(majors$semester) <- gsub("CURR", "Term ", levels(majors$semester))
 arr <- arrow(angle = 15, type = "closed", length = unit(.03, "npc"))
 ggplot(majors,
        aes(x = semester, stratum = curriculum, alluvium = student,
            fill = curriculum)) +
-  scale_fill_brewer(type = "qual", palette = "Set2") +
+  scale_fill_brewer(type = "qual", palette = "Set2", name = "Curriculum") +
   scale_y_continuous(breaks = seq(-2, 10, 2)) +
   geom_flow(stat = "alluvium", lode.guidance = "frontback",
             color = "darkgray", negate.strata = NA) +
   geom_stratum(negate.strata = NA) +
   theme(legend.position = "right") +
-  labs(x = NULL) +
+  labs(x = NULL, y = "Number of students") +
   # annotate axes
   annotate("segment", x = 2.33, y = -1.33, xend = 2, yend = -.33, arrow = arr) +
   annotate("segment", x = 3, y = -1.33, xend = 3, yend = -.33, arrow = arr) +
