### 2020-03-24 #2017 @corybrunson
@akeshavan i wonder if it's too late for me to move the file `paper.md` into the `paper` subdirectory. This would clean up the main directory a bit. But i gather it would affect whedon's ability to generate the PDF.
