### 2020-03-29 cfa7766
update documentation & website + upgrade version for cran release
