+++ man/vaccinations.Rd
@@ -3,7 +3,9 @@
 \name{vaccinations}
 \alias{vaccinations}
 \title{Influenza vaccination rates}
-\format{An alluvial data frame in lodes form.}
+\format{
+An alluvial data frame in lodes form.
+}
 \source{
 RAND American Life Panel \url{https://alpdata.rand.org/}
 }
