+++ DESCRIPTION
@@ -1,23 +1,22 @@
 Package: ggalluvial
 Type: Package
 Title: Alluvial Plots in 'ggplot2'
-Version: 0.11.2
-Date: 2019-12-03
+Version: 0.11.3
+Date: 2020-03-29
 Authors@R: person('Jason Cory', 'Brunson', email = 'cornelioid@gmail.com',
     role = c('aut', 'cre'))
 Maintainer: Jason Cory Brunson <cornelioid@gmail.com>
-Description: Alluvial plots use
-    variable-width ribbons and stacked bar plots to represent
-    multi-dimensional or repeated-measures data with categorical or ordinal
-    variables;
-    see Riehmann, Hanfler, and Froehlich (2005)
+Description: Alluvial plots use variable-width ribbons and stacked bar plots to
+    represent multi-dimensional or repeated-measures data with categorical or
+    ordinal variables; see Riehmann, Hanfler, and Froehlich (2005)
     <doi:10.1109/INFVIS.2005.1532152> and Rosvall and Bergstrom (2010)
     <doi:10.1371/journal.pone.0008694>.
-    Alluvial plots are statistical graphics in the sense of Wilkinson (2005)
-    <doi:>; they share elements with Sankey diagrams and parallel sets plots
-    but are uniquely determined from the data and a small set of parameters.
-    This package extends Wickham's () <doi:> layered grammar of graphics to
-    generate alluvial plots from tidy data.
+    Alluvial plots are statistical graphics in the sense of Wilkinson (2006)
+    <doi:10.1007/0-387-28695-0>; they share elements with Sankey diagrams and
+    parallel sets plots but are uniquely determined from the data and a small
+    set of parameters. This package extends Wickham's (2010)
+    <doi:10.1198/jcgs.2009.07098> layered grammar of graphics to generate
+    alluvial plots from tidy data.
 Depends:
     R (>= 3.3),
     ggplot2 (>= 2.2)
@@ -42,5 +41,5 @@ License: GPL-3 + file LICENSE
 URL: http://corybrunson.github.io/ggalluvial/
 BugReports: https://github.com/corybrunson/ggalluvial/issues
 VignetteBuilder: knitr
-RoxygenNote: 7.0.2
+RoxygenNote: 7.1.0
 Roxygen: list(markdown = TRUE)
