### 2020-04-03 94f551c
revise lode-ordering param to allow partial orderings + apply reverse to lode-ordering - resolves #52
