+++ R/stat-alluvium.r
@@ -30,8 +30,10 @@
 #'   (each of length the number of rows of `data`) or NULL entries (indicating
 #'   no imposed ordering), or else a numeric matrix of corresponding dimensions,
 #'   giving the preferred ordering of alluvia at each axis. This will be used to
-#'   order the lodes within each stratum by sorting the lodes first by stratum
-#'   and then by the provided vectors.
+#'   order the lodes within each stratum by sorting the lodes first by stratum,
+#'   then by the provided vectors, and lastly by remaining factors (if the
+#'   vectors contain duplicate entries and therefore do not completely determine
+#'   the lode orderings).
 #' @example inst/examples/ex-stat-alluvium.r
 #' @export
 stat_alluvium <- function(mapping = NULL,
@@ -219,42 +221,50 @@ StatAlluvium <- ggproto(
     # define 'deposit' variable to rank strata vertically
     data <- deposit_data(data, decreasing, reverse, absolute)
     
-    # if `lode.ordering` not provided, calculate it (ignoring aesthetics)
-    if (is.null(lode.ordering)) {
-      # invoke surrounding axes in the order prescribed by `lode.guidance`
-      if (is.character(lode.guidance)) {
-        lode.guidance <- get(paste0("lode_", lode.guidance))
-      }
-      stopifnot(is.function(lode.guidance))
-      # summary data of alluvial deposits
-      alluv_dep <- alluviate(data, "x", "deposit", "alluvium")
-      # axis indices
-      alluv_x <- setdiff(names(alluv_dep), "alluvium")
-      # calculate `lode.ordering` from `lode.guidance`
-      lode.ordering <- matrix(NA_integer_,
-                              nrow = nrow(alluv_dep), ncol = length(alluv_x))
-      dimnames(lode.ordering) <- list(alluv_dep$alluvium, alluv_x)
-      for (xx in alluv_x) {
-        ord_x <- lode.guidance(length(alluv_x), match(xx, alluv_x))
-        # order by aesthetics in order
-        lode.ordering[, xx] <- interaction(alluv_dep[, alluv_x[rev(ord_x)]],
-                                           drop = TRUE)
-      }
+    # invoke surrounding axes in the order prescribed by `lode.guidance`
+    if (is.character(lode.guidance)) {
+      lode.guidance <- get(paste0("lode_", lode.guidance))
+    }
+    stopifnot(is.function(lode.guidance))
+    # summary data of alluvial deposits
+    alluv_dep <- alluviate(data, "x", "deposit", "alluvium")
+    # axis indices
+    alluv_x <- setdiff(names(alluv_dep), "alluvium")
+    # ensure that `lode.ordering` is a matrix with column names
+    if (! is.null(lode.ordering)) {
       # bind a vector to itself to create a matrix
-    } else if (is.vector(lode.ordering)) {
-      lode.ordering <- matrix(lode.ordering,
-                              nrow = length(lode.ordering),
-                              ncol = length(unique(data$x)))
+      if (is.vector(lode.ordering)) {
+        lode.ordering <- matrix(lode.ordering,
+                                nrow = length(lode.ordering),
+                                ncol = length(unique(data$x)))
+      }
+      colnames(lode.ordering) <- alluv_x
+    }
+    # calculate `lode_ords` from `lode.guidance` and `lode.ordering`
+    lode_ords <- matrix(NA_integer_,
+                        nrow = nrow(alluv_dep), ncol = length(alluv_x))
[…]
