+++ R/alluvial-data.r
@@ -35,9 +35,6 @@
 #' @import tidyselect
 #' @family alluvial data manipulation
 #' @param data A data frame.
-#' @param logical Defunct. Whether to return a logical value or a character
-#'   string indicating the type of alluvial structure ("none", "lodes", or
-#'   "alluvia").
 #' @param silent Whether to print messages.
 #' @param key,value,id In `to_lodes_form`, handled as in
 #'   [tidyr::gather()] and used to name the new axis (key), stratum
@@ -51,14 +48,15 @@
 #' @param ... Used in `is_alluvia_form` and `to_lodes_form` as in
 #'   [dplyr::select()] to determine axis variables, as an alternative
 #'   to `axes`. Ignored when `axes` is provided.
-#' @param weight Optional field of `data`, handled using
-#'   [rlang::enquo()], to be used as heights or depths of the alluvia
-#'   or lodes.
-#' @param diffuse Fields of `data`, handled using
-#'   [tidyselect::vars_select()], to merge into the reshapen data by
-#'   `id`. They must be a subset of the axis variables. Alternatively, a
-#'   logical value indicating whether to merge all (`TRUE`) or none
-#'   (`FALSE`) of the axis variables.
+#' @param y Optional field(s) of `data`, handled using [enquo()], to be used as
+#'   heights or depths of the alluvia or lodes.
+#' @param y_to A string specifying the name of the column to create from the
+#'   column(s) identified by `y`. If needed for multiple columns but not
+#'   provided, defaults to 'y'.
+#' @param diffuse Fields of `data`, handled using [tidyselect::vars_select()],
+#'   to merge into the reshapen data by `id`. They must be a subset of the axis
+#'   variables. Alternatively, a logical value indicating whether to merge all
+#'   (`TRUE`) or none (`FALSE`) of the axis variables.
 #' @param distill A logical value indicating whether to include variables, other
 #'   than those passed to `key` and `value`, that vary within values
 #'   of `id`. Alternatively, a function (or its name) to be used to distill
@@ -76,13 +74,52 @@
 #' @export
 is_lodes_form <- function(data,
                           key, value, id,
-                          weight = NULL,
-                          logical = TRUE, silent = FALSE) {
+                          y = NULL,
+                          silent = FALSE) {
+  
+  key_var <- vars_pull(names(data), !! enquo(key))
+  value_var <- vars_pull(names(data), !! enquo(value))
+  id_var <- vars_pull(names(data), !! enquo(id))
+  if (key_var %in% c(value_var, id_var)) {
+    stop("`key` must be distinct from `id` and `value` variables.")
+  }
+  
+  if (any(duplicated(cbind(data[c(id_var, key_var)])))) {
+    if (! silent) warning("Some id-key (alluvium-axis) pairs are duplicated.")
+  }
+  
+  n_pairs <- dplyr::n_distinct(data[key_var]) * dplyr::n_distinct(data[id_var])
+  if (nrow(data) < n_pairs) {
+    if (! silent) warning("Some id-key (alluvium-axis) pairs are missing.")
+  }
+  
+  # if `y` is not `NULL`, use non-standard evaluation to identify `y_var`
+  if (! is.null(rlang::enexpr(y))) {
+    y_var <- vars_select(unique(names(data)), !! enquo(y))
+    if (length(y_var) > 0) {
+      if (length(y_var) > 1) stop("`y` must be a single variable.")
+      if (y_var %in% c(key_var, value_var, id_var)) {
+        stop("`y` must be distinct from id-key-value variables.")
+      }
[…]
