### 2020-04-07 454e5ea
documentation template for alt curves
