### 2020-04-09 3980cfe
debug knot.pos + add more fragile examples + update news
