+++ R/geom-flow.r
@@ -78,7 +78,7 @@ GeomFlow <- ggproto(
     }
 
     knot.pos <- params$knot.pos
-    if (is.null(knot.pos)) knot.pos <- 1/6
+    if (is.null(knot.pos)) knot.pos <- 1/4
 
     # positioning parameters
     transform(data,
@@ -96,8 +96,9 @@ GeomFlow <- ggproto(
     data <- data[complete.cases(data), ]
 
     # adjoin data with itself by alluvia along adjacent axes
-    flow_pos <- intersect(names(data), c("x", "xmin", "xmax", "width",
-                                         "y", "ymin", "ymax", "knot.pos"))
+    flow_pos <- intersect(names(data), c("x", "xmin", "xmax",
+                                         "width", "knot.pos",
+                                         "y", "ymin", "ymax"))
     flow_aes <- intersect(names(data), c("size", "linetype",
                                          "colour", "fill", "alpha"))
     flow_fore <- if (aes.flow != "backward") flow_aes else NULL
