+++ inst/examples/ex-geom-flow.r
@@ -28,6 +28,10 @@ wph$Year <- as.integer(as.character(wph$Year))
 ggplot(wph,
        aes(x = Year, alluvium = Region, y = Telephones)) +
   geom_flow(aes(fill = Region, colour = Region), width = 0)
+# hold the knot positions fixed
+ggplot(wph,
+       aes(x = Year, alluvium = Region, y = Telephones)) +
+  geom_flow(aes(fill = Region, colour = Region), width = 0, knot.fix = TRUE)
 
 # rightward flow aesthetics for vaccine survey data, with cubic flows
 data(vaccinations)
