### 2020-04-09 aa63d83
change knot-fix to knot-prop + add knot-prop tests
