### 2020-04-10 bbe5491
rm weight from documentation on y
