### 2020-04-11 eb6f0a1
do not negate alluvial order + update example
