### 2020-04-11 9978193
simplify error tests + update vdiffr tests
