### 2020-04-11 62400fd
update news re alluvia in negative strata
