+++ R/alluvial-data.r
@@ -1,55 +1,70 @@
-#' Check for alluvial structure and convert between alluvial formats
+#' @title Test for alluvial structure and convert between alluvial formats
+#'
+#' @description
+#'
+#' The functions `is_*_form()` determine whether a data frame is alluvial---that
+#' it has either the id--key--value structure of lodes format or the tidy
+#' properties of alluvia format. The functions `to_*_form()` convert data from
+#' one format to the other.
+#' 
+
+#' @details
 #'
 #' Alluvial plots consist of multiple horizontally-distributed columns (axes)
-#' representing factor variables, vertical divisions (strata) of these axes
-#' representing these variables' values; and splines (alluvial flows) connecting
-#' vertical subdivisions (lodes) within strata of adjacent axes representing
+#' representing categorical variables, vertical divisions (strata) of these axes
+#' representing these variables' values; and ribbons (alluvial flows) connecting
+#' vertical subdivisions (lodes) within strata of adjacent axes, representing
 #' subsets or amounts of observations that take the corresponding values of the
-#' corresponding variables. This function checks a data frame for either of two
-#' types of alluvial structure:
+#' corresponding variables. **ggalluvial** can generate an alluvial plot from
+#' data in either of two recognized formats:
+#' 
+
+#' - **Lodes format** encodes _one row per measurement_ or _one row per lode_,
+#' using an `id`--`key`--`value` structure: Each row contains the recorded
+#' `value` of one measurement (the `key`) for one subject or cohort (the `id`).
+#' Visually, the lode is the unique intersection of one alluvium (the `id`) and
+#' one axis (the `key`) and belongs to a unique stratum (the `value`).
+#' Additional columns may contain lode-level variables, including magnitudes and
+#' weights.
+
+#' - **Alluvia format** encodes _one row per subject_ or _one row per alluvium_,
+#' using a [tidy](https://tidyr.tidyverse.org/) structure: Each subject (or
+#' cohort) has a row, each measurement has a column, and each measured value is
+#' an entry. Additional columns may contain subject-level variables, including
+#' magnitudes and weights; but lode-level magnitudes and weights may also be
+#' encoded as column ranges having the same cardinality as those of the
+#' measurements.
+
 #'
-#' - One row per **lode**, wherein each row encodes a subset or amount of
-#'   observations having a specific profile of axis values, a `key` field
-#'   encodes the axis, a `value` field encodes the value within each axis, and a
-#'   `id` column identifies multiple lodes corresponding to the same subset or
-#'   amount of observations. `is_lodes_form` tests for this structure.
-#' - One row per **alluvium**, wherein each row encodes a subset or amount of
-#'   observations having a specific profile of axis values and a set `axes` of
-#'   fields encodes its values at each axis variable. `is_alluvia_form` tests
-#'   for this structure.
+#' The functions `is_*_form()` test for these structures. Both formats may
+#' include missing values, by omission (lodes format) or as `NA`s (both). See
+#' the primary vignette for several illustrations: `vignette("ggalluvial)`.
 #'
-#' `to_lodes_form` takes a data frame with several designated variables to
-#' be used as axes in an alluvial plot, and reshapes the data frame so that
[…]
