+++ man/alluvial-data.Rd
@@ -75,23 +75,23 @@ of the columns to contain alluvia (identifiers), axes (keys), and strata
 column(s) identified by \code{y}. If needed for multiple columns but not
 provided, defaults to 'y'.}
 
-\item{diffuse}{Fields of \code{data}, handled using \code{\link[tidyselect:vars_select]{tidyselect::vars_select()}},
-to merge into the reshapen data by \code{id}. They must be a subset of the axis
-variables. Alternatively, a logical value indicating whether to merge all
-(\code{TRUE}) or none (\code{FALSE}) of the axis variables.}
+\item{diffuse}{Fields of \code{data} to merge into the lengthened data, joining by
+\code{alluvia_to}. They must be among the variables passed to \code{axes}.
+Alternatively, a logical value indicating whether to merge all (\code{TRUE}) or
+none (\code{FALSE}) of these variables.}
 
 \item{discern}{Logical value indicating whether to suffix values of the
-variables used as axes that appear at more than one variable in order to
+variables passed to \code{axes} that appear at more than one axis, in order to
 distinguish their factor levels. This forces the levels of the combined
-factor variable \code{value} to be in the order of the axes.}
+factor variable \code{strata_to} to be in the order of the axes.}
 
-\item{distill}{A logical value indicating whether to include variables, other
-than those passed to \code{key} and \code{value}, that vary within values
-of \code{id}. Alternatively, a function (or its name) to be used to distill
-each such variable to a single value. In addition to existing functions,
-\code{distill} accepts the character values \code{"first"} (used if
-\code{distill} is \code{TRUE}), \code{"last"}, and \code{"most"} (which
-returns the first modal value).}
+\item{distill}{A logical value indicating whether variables excluded from
+\code{axes_from} and \code{strata_from} that vary within values of \code{alluvia_from}
+should be included in the result. Alternatively, a function (or its name)
+to be used to distill each such variable to a single value. In addition to
+existing functions, \code{distill} accepts the character values \code{"first"} (used
+if \code{distill} is \code{TRUE}), \code{"last"}, and \code{"most"} (which returns the first
+modal value).}
 }
 \description{
 The functions \verb{is_*_form()} determine whether a data frame is alluvial---that
