### 2020-04-14 ae133bd
change computed variable contact to position & nudge labels apart in order-rectangles vignette
