### 2020-04-14 b6ad6aa
document computed variable position
