+++ man-roxygen/computed-variables.r
@@ -6,7 +6,10 @@
 #'   \item{count}{cumulative weight of lode}
 #'   \item{prop}{weighted proportion of lode}
 #'   \item{stratum}{value of variable used to define strata}
-#'   \item{lode}{lode label distilled from alluvia}
+#'   \item{lode}{lode label distilled from alluvia
+#'               (`stat_alluvium()` and `stat_flow()` only)}
+#'   \item{position}{`"front"` or `"back"` of each flow
+#'                   (`stat_flow()` only)}
 #' }
 #' The numerical variables are grouped by `x` and weighted by `weight`. The
 #' character variable `lode` is obtained from `alluvium` according to `distill`.
