+++ R/stat-alluvium.r
@@ -23,14 +23,15 @@
 #'   ordering the lodes within each stratum, or else a character string
 #'   identifying the function. Character options are "zigzag", "frontback",
 #'   "backfront", "forward", and "backward" (see [`lode-guidance-functions`]).
-#' @param lode.ordering A list (of length the number of axes) of integer vectors
-#'   (each of length the number of rows of `data`) or NULL entries (indicating
-#'   no imposed ordering), or else a numeric matrix of corresponding dimensions,
-#'   giving the preferred ordering of alluvia at each axis. This will be used to
-#'   order the lodes within each stratum by sorting the lodes first by stratum,
-#'   then by the provided vectors, and lastly by remaining factors (if the
-#'   vectors contain duplicate entries and therefore do not completely determine
-#'   the lode orderings).
+#' @param lode.ordering **Deprecated in favor of the `order` aesthetic.** A list
+#'   (of length the number of axes) of integer vectors (each of length the
+#'   number of rows of `data`) or NULL entries (indicating no imposed ordering),
+#'   or else a numeric matrix of corresponding dimensions, giving the preferred
+#'   ordering of alluvia at each axis. This will be used to order the lodes
+#'   within each stratum by sorting the lodes first by stratum, then by the
+#'   provided vectors, and lastly by remaining factors (if the vectors contain
+#'   duplicate entries and therefore do not completely determine the lode
+#'   orderings).
 #' @example inst/examples/ex-stat-alluvium.r
 #' @export
 stat_alluvium <- function(mapping = NULL,
@@ -45,7 +46,7 @@ stat_alluvium <- function(mapping = NULL,
                           aggregate.y = NULL,
                           cement.alluvia = ggalluvial_opt("cement.alluvia"),
                           lode.guidance = ggalluvial_opt("lode.guidance"),
-                          lode.ordering = ggalluvial_opt("lode.ordering"),
+                          lode.ordering = NULL,
                           aes.bind = ggalluvial_opt("aes.bind"),
                           infer.label = FALSE,
                           min.y = NULL, max.y = NULL,
@@ -169,7 +170,7 @@ StatAlluvium <- ggproto(
                            aggregate.y = NULL,
                            cement.alluvia = ggalluvial_opt("cement.alluvia"),
                            lode.guidance = ggalluvial_opt("lode.guidance"),
-                           lode.ordering = ggalluvial_opt("lode.ordering"),
+                           lode.ordering = NULL,
                            aes.bind = ggalluvial_opt("aes.bind"),
                            infer.label = FALSE,
                            min.y = NULL, max.y = NULL) {
@@ -186,13 +187,32 @@ StatAlluvium <- ggproto(
       }
     }
     
-    # aesthetics (in prescribed order)
-    aesthetics <- intersect(c(.color_diff_aesthetics, .text_aesthetics),
-                            names(data))
+    # ensure that `lode.ordering` is a matrix with column names
+    if (! is.null(lode.ordering)) {
+      deprecate_parameter("lode.ordering",
+                          msg = "Use the `order` aesthetic instead.")
+      if (is.null(data$order)) {
+        # bind a vector to itself to create a matrix
+        if (is.vector(lode.ordering)) {
+          lode.ordering <- matrix(lode.ordering,
+                                  nrow = length(lode.ordering),
[…]
