+++ inst/examples/ex-stat-alluvium.r
@@ -21,15 +21,10 @@ gg + geom_flow(aes(fill = Survived, alpha = Sex), stat = "alluvium",
                lode.guidance = "forward")
 # prioritize aesthetic binding
 gg + geom_flow(aes(fill = Survived, alpha = Sex), stat = "alluvium",
-               aes.bind = TRUE, lode.guidance = "forward")
-# use of lode ordering
-lode_ord <- replicate(n = 3, expr = sample(x = 32), simplify = FALSE)
-print(lode_ord)
-gg + geom_flow(aes(fill = Survived, alpha = Sex), stat = "alluvium",
-               lode.ordering = lode_ord)
-# fixed lode ordering across axes
-gg + geom_flow(aes(fill = Survived, alpha = Sex), stat = "alluvium",
-               lode.ordering = lode_ord[[1]])
+               aes.bind = "alluvia", lode.guidance = "forward")
+# use of custom lode order
+gg + geom_flow(aes(fill = Survived, alpha = Sex, order = sample(x = 32)),
+               stat = "alluvium")
 # use of custom luide guidance function
 lode_custom <- function(n, i) {
   stopifnot(n == 3)
