### 2020-04-15 4f8cf3f
export guide-lodes function
