### 2020-04-16 133776a
fix links in README per CRAN issues
