### 2020-04-16 d8c5a6b
rm LICENSE file from DESCRIPTION + add to Rbuildignore
