+++ DESCRIPTION
@@ -37,7 +37,7 @@ Suggests:
     ggrepel,
     ggfittext (>= 0.6),
     vdiffr (>= 0.2)
-License: GPL-3 + file LICENSE
+License: GPL-3
 URL: http://corybrunson.github.io/ggalluvial/
 BugReports: https://github.com/corybrunson/ggalluvial/issues
 VignetteBuilder: knitr
