### 2020-04-23 4387233
simplify code
