### 2020-04-25 3cfd0e2
add NULL defaults to stratum & alluvium aesthetics
