+++ R/stat-alluvium.r
@@ -89,7 +89,8 @@ StatAlluvium <- ggproto(
   
   required_aes = c("x"),
   
-  default_aes = aes(weight = 1),
+  # `<new-aes> = NULL` prevents "unknown aesthetics" warnings
+  default_aes = aes(weight = 1, stratum = NULL, alluvium = NULL),
   
   setup_params = function(data, params) {
     
