### 2020-04-27 ff14a04
improve data documentation + lazy load data
