+++ inst/examples/ex-stat-alluvium.r
@@ -42,9 +42,8 @@ lode_custom <- function(n, i) {
 gg + geom_flow(aes(fill = Survived, alpha = Sex), stat = "alluvium",
                aes.bind = TRUE, lode.guidance = lode_custom)
 
-data(majors)
 # omit missing elements & reverse the `y` axis
-ggplot(majors,
+ggplot(ggalluvial::majors,
        aes(x = semester, stratum = curriculum, alluvium = student, y = 1)) +
   geom_alluvium(fill = "darkgrey", na.rm = TRUE) +
   geom_stratum(aes(fill = curriculum), color = NA, na.rm = TRUE) +
@@ -52,7 +51,7 @@ ggplot(majors,
   scale_y_reverse()
 
 # alluvium cementation examples
-gg <- ggplot(majors,
+gg <- ggplot(ggalluvial::majors,
              aes(x = semester, stratum = curriculum, alluvium = student,
                  fill = curriculum)) +
   geom_stratum()
