### 2020-04-27 f702dca
format computed-variables template
