+++ README.rmd
@@ -16,6 +16,7 @@ knitr::opts_chunk$set(
 
 [![Travis build status](https://travis-ci.org/corybrunson/ggalluvial.svg?branch=master)](https://travis-ci.org/corybrunson/ggalluvial)
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/ggalluvial)](https://cran.r-project.org/package=ggalluvial)
+[![JOSS](https://joss.theoj.org/papers/2a38fbb08d94994145d01c370f243723/status.svg)](https://joss.theoj.org/papers/2a38fbb08d94994145d01c370f243723)
 
 This is a [**ggplot2** extension](https://exts.ggplot2.tidyverse.org/) for alluvial plots.
 
