### 2020-05-20 #2017 @corybrunson
@arfon revisions during the review were only to the manuscript, and i happened to submit a necessary update to CRAN in the meantime (due to ggplot2 having discontinued dependence on plyr). Should i update the JOSS submission to this patch version? Either way, i'll plan to archive the current version on Zenodo as you've described.
