### 2020-05-20 #2017 @arfon
> Should i update the JOSS submission to this patch version? Either way, i'll plan to archive the current version on Zenodo as you've described.

Yes, I think that's fine.
