### 2020-05-20 dfb3ed8
docs(readme): shorten badge names + Zenodo DOI badge
